### 2019-07-29 #1604 @whedon
[PRE REVIEW]: Graph Transliterator: a graph-based transliteration tool
**Submitting author:** @seanpue (<a href="http://orcid.org/0000-0001-8463-8578">A. Sean Pue</a>)
**Repository:** <a href="https://github.com/seanpue/graphtransliterator" target ="_blank">https://github.com/seanpue/graphtransliterator</a>
**Version:** v0.2.7
