### 2019-07-29 #1604 @gkthiruvathukal
@kyleniemeyer I'm pretty sure I can handle it. Most of my assignments are moving toward a conclusion, so I am willing. After this, I might need to pause until late August as I will be on vacation starting August 10th with classes resuming on the 26th. Please go ahead and assign me.
