### 2019-07-30 33eea41
fixed max() problem in check_ambiguity if no rules, readme fix
