### 2019-07-30 afd00fa
fixed max() problem in check_ambiguity if no rules, readme fix
