### 2019-07-30 ed4d0da
added test for transliterator without rules
