### 2019-08-03 72f0161
used black, updated dev status, minor code changes
