+++ graphtransliterator/core.py
@@ -14,16 +14,22 @@ import yaml
 from .validate import validate_easyreading_settings, validate_settings
 from .process import _process_easyreading_settings
 from .initialize import (
-    _graph_from, _onmatch_rule_of, _onmatch_rules_lookup, _tokenizer_from,
-    _tokens_by_class_of, _tokens_of, _transliteration_rule_of, _whitespace_of
+    _graph_from,
+    _onmatch_rule_of,
+    _onmatch_rules_lookup,
+    _tokenizer_from,
+    _tokens_by_class_of,
+    _tokens_of,
+    _transliteration_rule_of,
+    _whitespace_of,
 )
 from .exceptions import (
     AmbiguousTransliterationRulesException,
     NoMatchingTransliterationRuleException,
-    UnrecognizableInputTokenException
+    UnrecognizableInputTokenException,
 )
 
-logger = logging.getLogger('graphtransliterator')
+logger = logging.getLogger("graphtransliterator")
 
 
 class GraphTransliterator:
@@ -162,38 +168,45 @@ class GraphTransliterator:
     from_dict : constructor from  dictionary in "easy reading" format
     from_yaml : constructor from YAML string in "easy reading" format
     from_yaml_file : constructor from YAML in "easy reading" format
-""" # noqa
+"""  # noqa
 
-    def __init__(self, tokens, rules, onmatch_rules, whitespace, metadata,
-                 check_settings=True, check_ambiguity=True,
-                 ignore_errors=False, **kwargs):
+    def __init__(
+        self,
+        tokens,
+        rules,
+        onmatch_rules,
+        whitespace,
+        metadata,
+        check_settings=True,
+        check_ambiguity=True,
+        ignore_errors=False,
+        **kwargs
+    ):
         if check_settings:
-            validate_settings(
-                tokens, rules, onmatch_rules, whitespace, metadata
-            )
+            validate_settings(tokens, rules, onmatch_rules, whitespace, metadata)
         self._metadata = metadata
         self._tokens = _tokens_of(tokens)
         self._tokens_by_class = _tokens_by_class_of(tokens)
         self._rules = sorted(
             [_transliteration_rule_of(rule) for rule in rules],
-            key=lambda transliteration_rule: transliteration_rule.cost)
+            key=lambda transliteration_rule: transliteration_rule.cost,
+        )
         if check_ambiguity:
             self.check_for_ambiguity()
         self._onmatch_rules = [_onmatch_rule_of(_) for _ in onmatch_rules]
-        self._onmatch_rules_lookup = \
-            _onmatch_rules_lookup(tokens, self._onmatch_rules)
+        self._onmatch_rules_lookup = _onmatch_rules_lookup(tokens, self._onmatch_rules)
         self._whitespace = _whitespace_of(whitespace)
         self._ignore_errors = ignore_errors
-        self.__init_parameters__ = {     # used by pruned_of
-            'tokens': tokens,
-            'rules': rules,
-            'onmatch_rules': onmatch_rules,
-            'whitespace': whitespace,
-            'metadata': metadata,
-            'check_settings': check_settings,
-            'ignore_errors': ignore_errors,
-            'check_ambiguity': check_ambiguity,
-            'kwargs': kwargs
+        self.__init_parameters__ = {  # used by pruned_of
+            "tokens": tokens,
+            "rules": rules,
+            "onmatch_rules": onmatch_rules,
+            "whitespace": whitespace,
+            "metadata": metadata,
+            "check_settings": check_settings,
+            "ignore_errors": ignore_errors,
+            "check_ambiguity": check_ambiguity,
+            "kwargs": kwargs,
         }
         self._tokenizer = _tokenizer_from(list(tokens.keys()))
[…]
