+++ graphtransliterator/initialize.py
@@ -18,8 +18,6 @@ def _tokens_of(tokens):
 
     return {key: set(value) for key, value in tokens.items()}
 
-# ---------- initialize tokens ----------
-
 
 def _tokens_by_class_of(tokens):
     """Generates lookup table of tokens in each class."""
@@ -40,13 +38,14 @@ def _transliteration_rule_of(rule):
     _cost = _cost_of(rule)
 
     transliteration_rule = TransliterationRule(
-        production=rule.get('production'),
-        prev_classes=rule.get('prev_classes'),
-        prev_tokens=rule.get('prev_tokens'),
-        tokens=rule.get('tokens'),
-        next_tokens=rule.get('next_tokens'),
-        next_classes=rule.get('next_classes'),
-        cost=_cost)
+        production=rule.get("production"),
+        prev_classes=rule.get("prev_classes"),
+        prev_tokens=rule.get("prev_tokens"),
+        tokens=rule.get("tokens"),
+        next_tokens=rule.get("next_tokens"),
+        next_classes=rule.get("next_classes"),
+        cost=_cost,
+    )
 
     return transliteration_rule
 
@@ -54,12 +53,13 @@ def _transliteration_rule_of(rule):
 def _num_tokens_of(rule):
     """Calculate the total number of tokens in a rule."""
 
-    total = 0
-    total += len(rule.get('prev_classes', []))
-    total += len(rule.get('prev_tokens', []))
-    total += len(rule.get('tokens'))
-    total += len(rule.get('next_tokens', []))
-    total += len(rule.get('next_classes', []))
+    total = (
+        len(rule.get("prev_classes", []))
+        + len(rule.get("prev_tokens", []))
+        + len(rule.get("tokens"))
+        + len(rule.get("next_tokens", []))
+        + len(rule.get("next_classes", []))
+    )
 
     return total
 
@@ -70,15 +70,17 @@ def _cost_of(rule):
     Rules requiring more tokens to match are made less costly and tried first.
     """
 
-    return math.log2(1+1/(1+_num_tokens_of(rule)))
+    return math.log2(1 + 1 / (1 + _num_tokens_of(rule)))
 
 
 def _onmatch_rule_of(onmatch_rule):
     """ Converts onmatch_rule into OnMatchRule. """
 
-    return OnMatchRule(prev_classes=onmatch_rule['prev_classes'],
-                       next_classes=onmatch_rule['next_classes'],
-                       production=onmatch_rule['production'])
+    return OnMatchRule(
+        prev_classes=onmatch_rule["prev_classes"],
+        next_classes=onmatch_rule["next_classes"],
+        production=onmatch_rule["production"],
+    )
 
 
 def _onmatch_rules_lookup(tokens, onmatch_rules):
@@ -101,7 +103,7 @@ def _onmatch_rules_lookup(tokens, onmatch_rules):
         for token_key, token_classes in tokens.items():
             # if the onmatch rule's next is of that class
             if rule.next_classes[0] in token_classes:
-                # interate through all tokens again
+                # iterate through all tokens again
                 for prev_token_key, prev_token_classes in tokens.items():
                     # if second token is of class of onmatch rule's last prev
                     # add the rule to curr -> prev
@@ -115,14 +117,17 @@ def _onmatch_rules_lookup(tokens, onmatch_rules):
                         rule_list.append(rule_i)
     return onmatch_lookup
 
+
 # ---------- initialize whitespace ---------
 
 
 def _whitespace_of(whitespace):
     """ Converts whitespace into WhiteSpace """
-    return WhitespaceRules(default=whitespace['default'],
-                           token_class=whitespace['token_class'],
[…]
