+++ tests/test_graphtransliterator.py
@@ -2,25 +2,21 @@
 # -*- coding: utf-8 -*-
 
 """Tests for `graphtransliterator` package."""
-
+import pkg_resources
 import pytest
 import yaml
-import graphtransliterator
 from graphtransliterator import GraphTransliterator
 from graphtransliterator.core import _easyreading_rule
 from graphtransliterator import process, validate
-from graphtransliterator.rules import (
-    OnMatchRule, TransliterationRule, WhitespaceRules
-)
-from graphtransliterator.graphs import (
-    DirectedGraph
-)
+from graphtransliterator.rules import OnMatchRule, TransliterationRule, WhitespaceRules
+from graphtransliterator.graphs import DirectedGraph
 
 from graphtransliterator.exceptions import (
     NoMatchingTransliterationRuleException,
     UnrecognizableInputTokenException,
-    AmbiguousTransliterationRulesException
+    AmbiguousTransliterationRulesException,
 )
+
 yaml_for_test = r"""
 tokens:
   a: [token, class1]
@@ -176,33 +172,33 @@ def test_graphtransliterator_process():
 
     data = yaml.safe_load(yaml_for_test)
 
-    assert process._process_rules({'a': 'A'})[0]['tokens'] == ['a']
-    assert process._process_rules({'a': 'A'})[0]['production'] == 'A'
-    assert process._process_onmatch_rules(
-        data['onmatch_rules'])[0]['prev_classes'][0] == 'class1'
-    assert process._process_onmatch_rules(
-        data['onmatch_rules'])[0]['next_classes'][0] == 'class2'
+    assert process._process_rules({"a": "A"})[0]["tokens"] == ["a"]
+    assert process._process_rules({"a": "A"})[0]["production"] == "A"
+    assert (
+        process._process_onmatch_rules(data["onmatch_rules"])[0]["prev_classes"][0]
+        == "class1"
+    )
+    assert (
+        process._process_onmatch_rules(data["onmatch_rules"])[0]["next_classes"][0]
+        == "class2"
+    )
 
 
 def test_graphtransliterator_models():
     """Test internal models."""
     tr = TransliterationRule(
-        production='A',
+        production="A",
         prev_classes=None,
         prev_tokens=None,
-        tokens=['a'],
+        tokens=["a"],
         next_tokens=None,
         next_classes=None,
-        cost=1
+        cost=1,
     )
     assert tr.cost == 1
     # assert TransliteratorOutput([tr], 'A').output == 'A'
-    assert OnMatchRule(prev_classes=['class1'],
-                       next_classes=['class2'],
-                       production=',')
-    assert WhitespaceRules(default=' ',
-                           token_class='wb',
-                           consolidate=False)
+    assert OnMatchRule(prev_classes=["class1"], next_classes=["class2"], production=",")
+    assert WhitespaceRules(default=" ", token_class="wb", consolidate=False)
 
 
 def test_graphtransliterator_structures():
@@ -213,11 +209,11 @@ def test_graphtransliterator_structures():
 
     graph.add_node({"type": "test1"})
     graph.add_node({"type": "test2"})
-    assert graph.node[0]['type'] == 'test1'
-    assert graph.node[1]['type'] == 'test2'
+    assert graph.node[0]["type"] == "test1"
+    assert graph.node[1]["type"] == "test2"
 
     graph.add_edge(0, 1, {"type": "edge_test1"})
-    assert graph.edge[0][1]['type'] == 'edge_test1'
+    assert graph.edge[0][1]["type"] == "edge_test1"
     assert type(graph.to_dict()) == dict
 
     # edge tail not in graph
[…]
