+++ HISTORY.rst
@@ -17,7 +17,7 @@ History
 
 0.2.10 (2019-08-03)
 -------------------
-* fixed test not working on travis
+* fixed test for version not working on travis
 
 0.2.9 (2019-08-03)
 ------------------
