### 2019-08-03 526b2f9
Bump version: 0.2.11 → 0.2.12
