### 2019-08-03 9a346be
fix to setup.cfg
