### 2019-08-03 bfa4a14
fixed version strings
