### 2019-08-03 b954f9c
added version tests
