### 2019-08-15 0c2a1bc
minor clean up
