### 2019-08-15 6ee3b08
removed yaml from validate
