+++ graphtransliterator/validate.py
@@ -1,49 +1,200 @@
 # -*- coding: utf-8 -*-
 """Methods and schemas to validate GraphTransliterator parameters."""
 
-from cerberus import Validator
+# from cerberus import Validator
+from collections import defaultdict
 from .process import RULE_RE, ONMATCH_RE
+
+from marshmallow import fields, Schema, validate, ValidationError, validates_schema
+
+# ----- Marshmallow schemas
+
+
+class WhitespaceSettingsSchema(Schema):
+    default = fields.Str(required=True)
+    token_class = fields.Str(required=True)
+    consolidate = fields.Boolean(required=True)
+
+
+class EasyReadingSettingsSchema(Schema):
+    tokens = fields.Dict(
+        keys=fields.Str(), values=fields.List(fields.Str()), required=True
+    )
+    rules = fields.Dict(
+        keys=fields.Str(validate=validate.Regexp(RULE_RE)),
+        values=fields.Str(),
+        required=True,
+    )
+    onmatch_rules = fields.List(
+        fields.Dict(
+            keys=fields.Str(validate=validate.Regexp(ONMATCH_RE)),
+            values=fields.Str(),
+            required=False,
+        )
+    )
+    metadata = fields.Dict(
+        keys=fields.Str(),
+        # no restriction on values
+        required=False,
+    )
+    whitespace = fields.Nested(WhitespaceSettingsSchema)
+
+
+class UserSchema(Schema):
+    name = fields.Str()
+    email = fields.Email()
+    created_at = fields.DateTime()
+
+
+class RuleSchema(Schema):
+    production = fields.Str()
+    tokens = fields.List(fields.Str())
+    prev_classes = fields.List(fields.Str())
+    next_classes = fields.List(fields.Str())
+    prev_tokens = fields.List(fields.Str())
+    next_tokens = fields.List(fields.Str())
+
+
+class OnMatchRuleSchema(Schema):
+    prev_classes = fields.List(fields.Str())
+    next_classes = fields.List(fields.Str())
+    production = fields.Str()
+
+
+class SettingsSchema(Schema):
+    tokens = fields.Dict(
+        keys=fields.Str(), values=fields.List(fields.Str()), required=True
+    )
+    rules = fields.Nested(RuleSchema, many=True, required=True)
+    whitespace = fields.Nested(WhitespaceSettingsSchema, many=False, required=True)
+    metadata = fields.Dict(
+        keys=fields.Str(), required=False  # no restriction on values
+    )
+    onmatch_rules = fields.Nested(OnMatchRuleSchema, many=True, required=False)
+    whitespace = fields.Nested(WhitespaceSettingsSchema, required=True)
+
+    from collections import defaultdict
+
+    @validates_schema
+    def validate_token_classes(self, data, **kwargs):
+        errors = defaultdict(list)
+        token_classes = list(set().union(*data["tokens"].values()))
+
+        # validate onmatch_rules and rules
+        for rule_type in ("onmatch_rules", "rules"):
+            for rule in data[rule_type]:
+                for component in ("prev_classes", "next_classes"):
+                    for _ in rule.get(component, []):
+                        if _ not in token_classes:
+                            errors[rule_type].append(
+                                'Invalid token class "{}" in {} of {}'.format(
+                                    _, component, rule
+                                )
+                            )
+
+        # validate whitespace token_class
+        whitespace = data["whitespace"]
+        whitespace_token_class = whitespace["token_class"]
+        if whitespace_token_class not in token_classes:
+            errors["whitespace"].append(
+                'Invalid token class "{}" in whitespace.'.format(
+                    whitespace_token_class, whitespace
+                )
+            )
+        if errors:
[…]
