### 2019-08-23 e729d15
fixed some serialization issues
