### 2019-08-23 7cc00d9
fixes to from_dict; revision of docs in process
