+++ graphtransliterator/core.py
@@ -100,9 +100,8 @@ class GraphTransliterator:
     Note
     ----
     This constructor does not validate settings and should typically not be called
-    directly unless constructing a GraphTransliterator programmatically. Use
-    :meth:`from_dict`, :meth:`from_easyreading_dict`, :meth:`from_yaml`, or
-    :meth:`from_yaml_file` instead for "easy reading" support.
+    directly. Use :meth:`from_dict` instead. For "easy reading" support, use
+    :meth:`from_easyreading_dict`, :meth:`from_yaml`, or :meth:`from_yaml_file` instead.
     Keyword parameters used here (``check_ambiguity``, ``ignore_errors``) can be passed
     from those other constructors.
 
@@ -110,6 +109,7 @@ class GraphTransliterator:
     ----------
     tokens : `dict` of {`str`: `set` of `str`}
         Mapping of input token types to token classes
+
     rules : `list` of `TransliterationRule`
         `list` of transliteration rules ordered by cost
 
@@ -129,7 +129,8 @@ class GraphTransliterator:
         exception. The default is false.
 
     check_ambiguity: `bool`, optional
-        If true (default), transliteration rules are checked for ambiguity.
+        If true (default), transliteration rules are checked for ambiguity. `load()`
+        and `load(s)` do not check ambiguity by default.
 
     onmatch_rules_lookup: `dict` of {`str`: dict o f {`str`: `list` of `int`}}, optional`
         OnMatchRules lookup, used internally, will be generated if not present.
@@ -145,7 +146,7 @@ class GraphTransliterator:
         not present.
 
     graphtransliterator_version: `str`, optional
-        Version of graphtransliterator used by `dump()` and `dumps()`.
+        Version of graphtransliterator, added by `dump()` and `dumps()`.
 
     Example
     -------
@@ -816,7 +817,21 @@ class GraphTransliterator:
         return self._whitespace
 
     @classmethod
-    def from_dict(cls, settings, **kwargs):
+    def from_dict(cls, dict_settings, **kwargs):
+        """Generate GraphTransliterator from `dict` settings.
+        Parameters
+        ----------
+        dict_settings : `dict`
+            Dictionary of settings
+
+        Returns
+        -------
+        graphtransliterator.GraphTransliterator
+            Graph transliterator
+
+
+        """
+        settings = SettingsSchema().load(dict_settings)
         return GraphTransliterator(
             settings["tokens"],
             settings["rules"],
@@ -906,10 +921,8 @@ class GraphTransliterator:
         _ = EasyReadingSettingsSchema().load(easyreading_settings)
         # convert to regular settings
         _ = _process_easyreading_settings(_)
-        # validate settings
-        settings = SettingsSchema().load(_)
-        # return GraphTransliterator
-        return GraphTransliterator.from_dict(settings, **kwargs)
+        # validation done in from_dict
+        return GraphTransliterator.from_dict(_, **kwargs)
 
     @classmethod
     def from_yaml(cls, yaml_str, charnames_escaped=True, **kwargs):
@@ -1020,8 +1033,8 @@ class GraphTransliterator:
         ... '''
         >>> gt = GraphTransliterator.from_yaml(yaml_)
         >>> gt.dumps()
[…]
