+++ graphtransliterator/core.py
@@ -43,6 +43,8 @@ logger = logging.getLogger("graphtransliterator")
 
 
 class GraphTransliteratorSchema(Schema):
+    """Schema for Graph Transliterator."""
+
     tokens = fields.Dict(
         keys=fields.Str(), values=fields.List(fields.Str()), required=True
     )
@@ -101,7 +103,7 @@ class GraphTransliterator:
     ----
     This constructor does not validate settings and should typically not be called
     directly. Use :meth:`from_dict` instead. For "easy reading" support, use
-    :meth:`from_easyreading_dict`, :meth:`from_yaml`, or :meth:`from_yaml_file` instead.
+    :meth:`from_easyreading_dict`, :meth:`from_yaml`, or :meth:`from_yaml_file`.
     Keyword parameters used here (``check_ambiguity``, ``ignore_errors``) can be passed
     from those other constructors.
 
@@ -150,71 +152,19 @@ class GraphTransliterator:
 
     Example
     -------
-    >>> from graphtransliterator import GraphTransliterator
-    >>> settings = {
-    ...    'tokens': {'a': set(['class1']),
-    ...               'b': set(['class2']),
-    ...               ' ': set(['wb']},
-    ...    'rules': [
-    ...        {'production': 'A', 'tokens': ['a']},
-    ...        {'production': 'B', 'tokens': ['b']},
-    ...        {'production': ' ', 'tokens': [' ']},
-    ...        {'production': 'A*',
-    ...         'prev_classes': ['class2'],
-    ...         'prev_tokens': ['a'],
-    ...         'tokens': ['a'],
-    ...         'next_tokens': ['a'],
-    ...         'next_classes': ['class2'] }],
-    ...    'onmatch_rules': [
-    ...        {'prev_classes': ['class1'],
-    ...         'next_classes': ['class1'],
-    ...         'production': ','}],
-    ...    'whitespace': {
-    ...        'default': ' ',
-    ...        'consolidate': False,
-    ...        'token_class': 'wb'},
-    ...    'metadata': {
-    ...        'author': 'Author McAuthorson',
-    ...        'version': '0.0.1'}
-    ... }
-    >>> gt = GraphTransliterator(
-    ...         settings['tokens'], settings['rules'], settings['onmatch_rules'],
-    ...         settings['whitespace'], settings['metadata']
-    ...      )
-    >>> gt.transliterate('baaab')
-    'BA,A*,AB'
-    >>>
+    >>> from graphtransliterator import *
+    >>> settings = {'tokens': {'a': {'vowel'}, ' ': {'wb'}}, 'onmatch_rules': [OnMatchRule(prev_classes=['vowel'], next_classes=['vowel'], production=',')], 'rules': [TransliterationRule(production='A', prev_classes=None, prev_tokens=None, tokens=['a'], next_tokens=None, next_classes=None, cost=0.5849625007211562), TransliterationRule(production=' ', prev_classes=None, prev_tokens=None, tokens=[' '], next_tokens=None, next_classes=None, cost=0.5849625007211562)], 'metadata': {'author': 'Author McAuthorson'}, 'whitespace': WhitespaceRules(default=' ', token_class='wb', consolidate=False)}
+    >>> gt = GraphTransliterator(**settings)
+    >>> gt.transliterate('a')
+    'A'
 
     See Also
     --------
-    from_easyreading_dict : constructor from  dictionary in "easy reading" format
-    from_yaml : constructor from YAML string in "easy reading" format
-    from_yaml_file : constructor from YAML in "easy reading" format
+    from_dict : Constructor from dictionary of settings
+    from_easyreading_dict : Constructor from  dictionary in "easy reading" format
+    from_yaml : Constructor from YAML string in "easy reading" format
[…]
