+++ requirements.txt
@@ -1,3 +1,2 @@
-cerberus==1.3.1
 pyyaml==5.1.2
 marshmallow==3.0.0rc9
