+++ graphtransliterator/core.py
@@ -115,7 +115,7 @@ class GraphTransliterator:
     rules : `list` of `TransliterationRule`
         `list` of transliteration rules ordered by cost
 
-    onmatch_rules : `list` of `OnMatchRule`, or `None`
+    onmatch_rules : `list` of :class:`OnMatchRule`, or `None`
         Rules for output to be inserted between tokens
         of certain classes when a transliteration rule has been matched
         but before its production string has been added to the output
@@ -131,10 +131,10 @@ class GraphTransliterator:
         exception. The default is false.
 
     check_ambiguity: `bool`, optional
-        If true (default), transliteration rules are checked for ambiguity. `load()`
-        and `load(s)` do not check ambiguity by default.
+        If true (default), transliteration rules are checked for ambiguity. :meth:`load()`
+        and :meth:`loads` do not check ambiguity by default.
 
-    onmatch_rules_lookup: `dict` of {`str`: dict o f {`str`: `list` of `int`}}, optional`
+    onmatch_rules_lookup: `dict` of {`str`: dict of {`str`: `list` of `int`}}, optional`
         OnMatchRules lookup, used internally, will be generated if not present.
 
     tokens_by_class: `dict` of {`str`: `set` of `str`}, optional
@@ -758,7 +758,7 @@ class GraphTransliterator:
 
     @property
     def onmatch_rules(self):
-        """`list` of `OnMatchRules`: Rules for productions between matches."""
+        """`list` of :class:`OnMatchRules`: Rules for productions between matches."""
         return self._onmatch_rules
 
     @property
@@ -769,6 +769,7 @@ class GraphTransliterator:
     @classmethod
     def from_dict(cls, dict_settings, **kwargs):
         """Generate GraphTransliterator from `dict` settings.
+
         Parameters
         ----------
         dict_settings : `dict`
@@ -776,10 +777,8 @@ class GraphTransliterator:
 
         Returns
         -------
-        graphtransliterator.GraphTransliterator
+        GraphTransliterator
             Graph transliterator
-
-
         """
         settings = SettingsSchema().load(dict_settings)
         return GraphTransliterator(
@@ -951,16 +950,11 @@ class GraphTransliterator:
     def dumps(self):
         """
         Dump settings of Graph Transliterator to Javascript Object Notation (JSON).
+
         Returns
         -------
         str
             JSON string
-
-        Note
-        ----
-        `OnmatchRule`, `TransliterationRule`, and `WhitespaceRules` are initialized from
-        :meth:`collections.namedtuple`. Their keys will always be in the same order and
-        can be accessed by index.
 
         Examples
         --------
@@ -983,7 +977,7 @@ class GraphTransliterator:
         ... '''
         >>> gt = GraphTransliterator.from_yaml(yaml_)
         >>> gt.dumps()
[…]
