+++ docs/usage.rst
@@ -10,11 +10,10 @@ To use Graph Transliterator in a project::
 
 Overview
 ========
-Graph Transliterator requires that you first configure
-a :class:`GraphTransliterator`. Then you can transliterate an input string
-using :meth:`GraphTransliterator.transliterate`. There are a few
-additional methods that can be used to extract information for specific use
-cases, such as details about which rules were matched.
+Graph Transliterator requires that you first configure a :class:`GraphTransliterator`.
+Then you can transliterate an input string using :meth:`transliterate`. There are a few
+additional methods that can be used to extract information for specific use cases, such
+as details about which rules were matched.
 
 Configuration
 =============
@@ -32,17 +31,16 @@ Graph Transliterator takes the following parameters:
 Initialization
 --------------
 
-Defining the rules for transliteration can be difficult, especially when
-dealing with complex scripts. That is why Graph Transliterator uses an
-"easy reading" format that allows you to enter the transliteration rules in
-the popular `YAML <https://yaml.org/>`_ format, either from a string
-(using :func:`GraphTransliterator.from_yaml`) or by reading from a file
-or stream (:func:`GraphTransliterator.from_yaml_file`). You can also
+Defining the rules for transliteration can be difficult, especially when dealing with
+complex scripts. That is why Graph Transliterator uses an "easy reading" format that
+allows you to enter the transliteration rules in the popular `YAML <https://yaml.org/>`_
+format, either from a string (using :func:`from_yaml`) or by reading
+from a file or stream (:func:`GraphTransliterator.from_yaml_file`). You can also
 initialize from the loaded contents of YAML
 (:func:`GraphTransliterator.from_easyreading_dict`).
 
-Here is a quick sample that parameterizes :class:`GraphTransliterator` using an
-easy reading YAML string (with comments):
+Here is a quick sample that parameterizes :class:`GraphTransliterator` using an easy
+reading YAML string (with comments):
 
 >>> from graphtransliterator import GraphTransliterator
 >>> yaml = """
@@ -68,22 +66,22 @@ easy reading YAML string (with comments):
 >>> gt_one.transliterate('aabb')
 '<TWO_As>B'
 
-The example above shows a very simple transliterator that replaces the
-input token "a" with "A", "bb" with "B", " " with " ", and two "a" in a row
-with "<2AS>". It does not consolidate whitespace, and treats " " as its
-default whitespace token. Tokens contain strings of one or more characters.
+The example above shows a very simple transliterator that replaces the input token "a"
+with "A", "bb" with "B", " " with " ", and two "a" in a row with "<2AS>". It does not
+consolidate whitespace, and treats " " as its default whitespace token. Tokens contain
+strings of one or more characters.
 
 Input Tokens and Token Class Settings
 -------------------------------------
-During transliteration, Graph Transliterator first attempts to convert the
-input string into a list of tokens. This is done internally using
+During transliteration, Graph Transliterator first attempts to convert the input string
+into a list of tokens. This is done internally using
 :meth:`GraphTransliterator.tokenize`:
 
   >>> gt_one.tokenize('abba')
   [' ', 'a', 'bb', 'a', ' ']
 
-Note that the default whitespace  token is added to the start and end of the
-input tokens.
+Note that the default whitespace  token is added to the start and end of the input
+tokens.
 
 Tokens can be more than one character, and longer tokens are matched first:
 
@@ -108,29 +106,29 @@ Tokens can be more than one character, and longer tokens are matched first:
 >>> gt_two.transliterate('aaa')
 '<DOUBLE_A><SINGLE_A>'
 
[…]
