### 2019-08-24 6c667c0
fixes to docs
