### 2019-08-24 2ca1ffa
doc updates
