+++ graphtransliterator/core.py
@@ -104,7 +104,7 @@ class GraphTransliterator:
     This constructor does not validate settings and should typically not be called
     directly. Use :meth:`from_dict` instead. For "easy reading" support, use
     :meth:`from_easyreading_dict`, :meth:`from_yaml`, or :meth:`from_yaml_file`.
-    Keyword parameters used here (``check_ambiguity``, ``ignore_errors``) can be passed
+    Keyword parameters used here (``ignore_errors``, ``check_ambiguity``) can be passed
     from those other constructors.
 
     Parameters
@@ -185,7 +185,7 @@ class GraphTransliterator:
         self._rules = rules
         self._tokens_by_class = tokens_by_class or _tokens_by_class_of(tokens)
 
-        self._check_ambiguity = check_ambiguity  # used by pruned_of
+        self._check_ambiguity = check_ambiguity
         if check_ambiguity:
             self.check_for_ambiguity()
 
@@ -562,7 +562,7 @@ class GraphTransliterator:
                 )
                 # No parsing rule was found at this location
                 if self.ignore_errors:
-                    # move along if ignoring errors
+                    # Move along if ignoring errors
                     token_i += 1
                     continue
                 else:
@@ -977,7 +977,7 @@ class GraphTransliterator:
         ... '''
         >>> gt = GraphTransliterator.from_yaml(yaml_)
         >>> gt.dumps()
-        '{"tokens": {"a": ["vowel"], " ": ["wb"]}, "rules": [{"production": "A", "prev_classes": null, "prev_tokens": null, "tokens": ["a"], "next_classes": null, "next_tokens": null, "cost": 0.5849625007211562}, {"production": " ", "prev_classes": null, "prev_tokens": null, "tokens": [" "], "next_classes": null, "next_tokens": null, "cost": 0.5849625007211562}], "whitespace": {"consolidate": false, "default": " ", "token_class": "wb"}, "onmatch_rules": [{"prev_classes": ["vowel"], "next_classes": ["vowel"], "production": ","}], "metadata": {"author": "Author McAuthorson"}, "onmatch_rules_lookup": {"a": {"a": [0]}}, "tokens_by_class": {"vowel": ["a"], "wb": [" "]}, "graph": {"node": [{"type": "Start", "ordered_children": {"a": [1], " ": [3]}}, {"type": "token", "token": "a", "ordered_children": {"__rules__": [2]}}, {"type": "rule", "rule_key": 0, "rule": {"production": "A", "prev_classes": null, "prev_tokens": null, "tokens": ["a"], "next_tokens": null, "next_classes": null, "cost": 0.5849625007211562}, "accepting": true, "ordered_children": {}}, {"type": "token", "token": " ", "ordered_children": {"__rules__": [4]}}, {"type": "rule", "rule_key": 1, "rule": {"production": " ", "prev_classes": null, "prev_tokens": null, "tokens": [" "], "next_tokens": null, "next_classes": null, "cost": 0.5849625007211562}, "accepting": true, "ordered_children": {}}], "edge": {"0": {"1": {"token": "a", "cost": 0.5849625007211562}, "3": {"token": " ", "cost": 0.5849625007211562}}, "1": {"2": {"cost": 0.5849625007211562}}, "3": {"4": {"cost": 0.5849625007211562}}}, "edge_list": [[0, 1], [1, 2], [0, 3], [3, 4]]}, "tokenizer_pattern": "(a|\\\\ )", "graphtransliterat
[…]
