### 2019-08-29 #1604 @seanpue
Hi @gkthiruvathukal Sorry for the delay. As an academic, I do not normally recommend reviewers. I also don't know many people on that list. I would suggest at least one reviewer with good Python, someone with some CS or basic graph theory background (particularly to check my descriptions), and also NLP, computational linguistics, or digital humanities. So quickly skimming that list, I would suggest desilinguist, mbod, rlskoeser, jkhan,richardlitt, linuxscout, alexhanna, or adammichaelwood. I will respond quickly to any further message. Thanks!
