### 2019-08-29 #1604 @gkthiruvathukal
Hi @seanpue, thanks for the response! I'm an academic here, too, but we use a slightly different model here at JOSS, especially since the review takes place in the open.  As reviewers start woring on your submission, they'll raise comments and issues, and you'll be able to respond to them.  Thanks for responding. I will see who is willing to review, and we'll go from there!
