### 2019-08-29 a3d6af8
minor revisions
