+++ README.rst
@@ -58,7 +58,9 @@ Features
 * Can provide **details** about each transliteration rule match
 * Allows **optional matching of all possible rules** in a particular location
 * Permits **pruning of rules** with certain productions
-* Can be **serialized** as a dictionary for export to JSON, etc.
+* **Validates**, as well as **serializes** to and **deserializes** from JSON
+  and Python data types, using accessible
+  `marshmallow <https://github.com/marshmallow-code/marshmallow>` schemas
 * Provides **full support for Unicode**, including Unicode **character names**
   in the "easy reading" YAML format
 * Constructs and uses a **directed tree** and performs a **best-first search**
