### 2019-08-29 958d1d2
more changes to setup.cfg
