+++ setup.cfg
@@ -12,8 +12,13 @@ search = __version__ = "{current_version}"
 replace = __version__ = "{new_version}"
 
 [bumpversion:file:graphtransliterator/core.py]
-search = 'graphtransliterator_version', '{current_version}'
-replace = 'graphtransliterator_version', '{new_version}'
+search = "graphtransliterator_version": "{current_version}"
+replace = "graphtransliterator_version": "{new_version}"
+
+[bumpversion:file:graphtransliterator/core.py]
+search = 'graphtransliterator_version': '{current_version}'
+replace = 'graphtransliterator_version': '{new_version}'
+
 universal = 1
 
 [flake8]
