### 2019-08-29 9c29dbb
updated to setup.cfg
