+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.3.0
+current_version = 0.3.1
 commit = True
 tag = True
 
@@ -23,3 +23,4 @@ test = pytest
 
 [tool:pytest]
 collect_ignore = ['setup.py']
+
