### 2019-08-30 2ec621f
fixes to readme
