### 2019-09-01 89dc914
Update azure-pipelines.yml for Azure Pipelines
