### 2019-09-01 fe5af1c
Update azure-pipelines.yml for Azure Pipelines
