+++ azure-pipelines.yml
@@ -25,10 +25,9 @@ steps:
 
 - script: |
     python -m pip install --upgrade pip
-    pip install -r requirements-dev.txt
+    pip install -r requirements_dev.txt
   displayName: 'Install dependencies'
 
 - script: |
-    pip install tox
     tox
   displayName: 'tox'
