### 2019-09-01 509e159
Update azure-pipelines.yml for Azure Pipelines
