### 2019-09-06 #1604 @seanpue
Hey @gkthiruvathukal Since @rlskoeser has the digital humanities, text-analysis, and nlp covered, perhaps you could ask people who do more general graphs and/or algorithms. From that list: hausen, gabrielspadon, juliohm, sonal-raj, vc1492a (also NLP, missed earlier), camillescott, szhorvat, amritagos.  Cheers.
