### 2019-09-06 #1604 @gkthiruvathukal
Thanks, @seanpue

@hausen, @gabrielspadon, @juliohm, @sonal-raj, @vc1492a, @camillescott, @szhorvat, @amritagos, are any of you willing to act as an additional reviewer for this submission, focused on the graph/algorithmic aspects? Once I have a second reviewer, I can commence review of @seanpue's submission.
