### 2019-09-07 #1604 @vc1492a
@gkthiruvathukal Thanks for the consideration, willing and happy to review this. Bandwidth has been a little tight as of late but this project is very interesting to me and I would enjoy the opportunity to perform and provide a review. You may want to add a third reviewer for NLP - I have some experience in this area but my experience is far from extensive.
