### 2019-09-07 #1717 @whedon
[REVIEW]: Graph Transliterator: a graph-based transliteration tool
**Submitting author:** @seanpue (<a href="http://orcid.org/0000-0001-8463-8578">A. Sean Pue</a>)
**Repository:** <a href="https://github.com/seanpue/graphtransliterator" target ="_blank">https://github.com/seanpue/graphtransliterator</a>
**Version:** v1.0.4
