### 2019-09-13 a63ad3d
package update and minor comment adjustments in core.py and schemas.py
