+++ setup.py
@@ -11,7 +11,7 @@ with open("README.rst") as readme_file:
 with open("HISTORY.rst") as history_file:
     history = history_file.read()
 
-requirements = ["cerberus", "pyyaml", "marshmallow"]
+requirements = ["marshmallow", "pyyaml"]
 
 setup_requirements = ["pytest-runner"]
 
