+++ HISTORY.rst
@@ -15,6 +15,12 @@ History
 [Unreleased-TODO]
 -----------------
 
+0.3.3 (2019-09-14)
+------------------
+* fixed missing marshmallow dependency (:pr:`47`). Thanks :user:`vc1492a`.
+* removed unused code from test (:pr:`47`). Thanks :user:`vc1492a`.
+* removed cerberus dependency.
+
 0.3.2 (2019-08-30)
 ------------------
 * fixed error in README.rst
