### 2019-09-15 6e0f802
added sphinx-issues and fixed documentation warnings
