+++ HISTORY.rst
@@ -15,6 +15,15 @@ History
 [Unreleased-TODO]
 -----------------
 
+#.#.# (####-##-##)
+------------------
+* added sphinx-issues to requirements_dev.txt
+* added sphinx_issues and settings to docs/conf.py
+* added GraphTransliteratorException import to __init__.py
+* added docs/_static directory
+* fixed emphasis error and duplicate object description in docs/usages.rst
+* fixed docstring in core.py
+
 0.3.3 (2019-09-14)
 ------------------
 * fixed missing marshmallow dependency (:pr:`47`). Thanks :user:`vc1492a`.
