+++ .readthedocs.yml
@@ -20,4 +20,4 @@ formats: all
 python:
   version: 3.7
   install:
-    - requirements: docs/requirements_dev.txt
+    - requirements: requirements_dev.txt
