### 2019-09-15 edb9713
modifications to authors, installation.rst, and added NOTICE
