+++ requirements.txt
@@ -1,2 +1,2 @@
-pyyaml==5.1.2
 marshmallow==3.0.5
+pyyaml==5.1.2
