### 2019-09-15 1d73888
removed history from setup.py
