+++ setup.py
@@ -8,9 +8,6 @@ from setuptools import setup, find_packages
 with open("README.rst") as readme_file:
     readme = readme_file.read()
 
-with open("HISTORY.rst") as history_file:
-    history = history_file.read()
-
 requirements = ["marshmallow", "pyyaml"]
 
 setup_requirements = ["pytest-runner"]
@@ -33,11 +30,13 @@ setup(
     description="A graph-based transliteration tool",
     install_requires=requirements,
     license="MIT license",
-    long_description=readme + "\n\n" + history,
+    long_description=readme,
+    long_description_content_type="text/x-rst",
     include_package_data=True,
     keywords="graphtransliterator",
     name="graphtransliterator",
     packages=find_packages(include=["graphtransliterator"]),
+    python_requires=">=3.5",
     setup_requires=setup_requirements,
     test_suite="tests",
     tests_require=test_requirements,
