### 2019-09-15 4ecbab7
make schemas compactable and added additional tests
