+++ graphtransliterator/core.py
@@ -31,7 +31,7 @@ from .schemas import (
     WhitespaceSettingsSchema,
 )
 from collections import deque
-from marshmallow import fields, post_load, Schema
+from marshmallow import fields, post_load, Schema, validates_schema, ValidationError
 
 logger = logging.getLogger("graphtransliterator")
 
@@ -50,11 +50,14 @@ class GraphTransliteratorSchema(Schema):
     metadata = fields.Dict(
         keys=fields.Str(), required=False  # No restriction on values
     )
-    onmatch_rules_lookup = fields.Dict()
-    tokens_by_class = fields.Dict(keys=fields.Str(), values=fields.List(fields.Str))
-    graph = fields.Nested(DirectedGraphSchema)
-    tokenizer_pattern = fields.Str()
-    graphtransliterator_version = fields.Str()
+    ignore_errors = fields.Bool(required=False)
+    onmatch_rules_lookup = fields.Dict(required=False)
+    tokens_by_class = fields.Dict(
+        keys=fields.Str(), values=fields.List(fields.Str), required=False
+    )
+    graph = fields.Nested(DirectedGraphSchema, required=False)
+    tokenizer_pattern = fields.Str(required=False)
+    graphtransliterator_version = fields.Str(required=False)
 
     class Meta:
         fields = (
@@ -63,6 +66,7 @@ class GraphTransliteratorSchema(Schema):
             "whitespace",
             "onmatch_rules",
             "metadata",
+            "ignore_errors",
             "onmatch_rules_lookup",
             "tokens_by_class",
             "graph",
@@ -79,6 +83,14 @@ class GraphTransliteratorSchema(Schema):
         # Do not check ambiguity if deserializing serialized GraphTransliterator
         data["check_ambiguity"] = False
         return GraphTransliterator(**data)
+
+    @validates_schema
+    def validate_onmatch_rules_lookup(self, data, **kwargs):
+        """Check that if there are onmatch_rules_lookup there are onmatch_rules."""
+        if data.get("onmatch_rules_lookup") and not data.get("onmatch_rules"):
+            raise ValidationError(
+                "Contains onmatch_rules_lookup but not onmatch_rules."
+            )
 
 
 class GraphTransliterator:
@@ -1038,26 +1050,27 @@ class GraphTransliterator:
 
         Example
         -------
-        >>> import graphtransliterator
+        >>> from graphtransliterator import GraphTransliterator
         >>> yaml_ = '''
-        ...       tokens:
-        ...         a: [vowel]
-        ...         ' ': [wb]
-        ...       rules:
-        ...         a: A
-        ...         ' ': ' '
-        ...       whitespace:
-        ...         default: " "
-        ...         consolidate: false
-        ...         token_class: wb
-        ...       onmatch_rules:
-        ...         - <vowel> + <vowel>: ','  # add a comma between vowels
-        ...       metadata:
-        ...         author: "Author McAuthorson"
-        ...     '''
-        >>> gt = graphtransliterator.GraphTransliterator.from_yaml(yaml_)
+        ... tokens:
+        ...   a: [vowel]
+        ...   ' ': [wb]
+        ... rules:
+        ...   a: A
+        ...   ' ': ' '
+        ... whitespace:
+        ...   default: " "
+        ...   consolidate: false
+        ...   token_class: wb
+        ... onmatch_rules:
+        ...   - <vowel> + <vowel>: ','  # add a comma between vowels
+        ... metadata:
+        ...   author: "Author McAuthorson"
+        ... '''
+        >>> gt = GraphTransliterator.from_yaml(yaml_)
         >>> gt.dump()
[…]
