### 2019-09-15 4253b51
added tests for all optional fields in load()
