### 2019-09-16 83b2777
updated docstrings in core.py
