+++ graphtransliterator/core.py
@@ -835,7 +835,8 @@ class GraphTransliterator:
 
         Returns
         -------
-        `GraphTransliterator`
+        GraphTransliterator
+            Graph Transliterator
 
         Note
         ----
@@ -988,7 +989,9 @@ class GraphTransliterator:
 
         See Also
         --------
-        dump : Dump GraphTransliterator configuration to Python types
+        dump : Dump Graph Transliterator configuration to Python data types
+        load : Load Graph Transliteration from configuration in Python data types
+        loads : Load Graph Transliteration from configuration as a JSON string
         """  # noqa
 
         return GraphTransliteratorSchema().dumps(self)
@@ -1004,7 +1007,7 @@ class GraphTransliterator:
 
                 ``"tokens"``
                   Mappings of tokens to their classes
-                  (`dict` of {str: `list` of `str`})
+                  (`OrderedDict` of {str: `list` of `str`})
 
                 ``"rules"``
                   Transliteration rules in direct format
@@ -1071,7 +1074,10 @@ class GraphTransliterator:
 
         See Also
         --------
-        dumps: Dump GraphTransliterator settings to Javascript Object Notation (JSON) str
+        dumps : Dump Graph Transliterator configuration to JSON string
+        load : Load Graph Transliteration from configuration in Python data types
+        loads : Load Graph Transliteration from configuration as a JSON string
+
 """  # noqa
         return GraphTransliteratorSchema().dump(self)
 
@@ -1090,7 +1096,7 @@ class GraphTransliterator:
 
                 ``"rules"``
                   Transliteration rules in direct format
-                  (`list` of `dict` of {`str`: `str`})
+                  (`list` of `OrderedDict` of {`str`: `str`})
 
                 ``"whitespace"``
                   Whitespace settings
@@ -1126,6 +1132,11 @@ class GraphTransliterator:
 
                 ``"graphtransliterator_version"``
                   Module version of `graphtransliterator` (`str`, optional)
+
+        Returns
+        -------
+        GraphTransliterator
+            Graph Transliterator
 
         Example
         -------
@@ -1218,12 +1229,41 @@ class GraphTransliterator:
         'A,A'
         >>>
 
+        See Also
+        --------
+        dump : Dump Graph Transliterator configuration to Python data types
+        dumps : Dump Graph Transliterator configuration to JSON string
+        loads : Load Graph Transliteration from configuration as a JSON string
         """  # noqa
         return GraphTransliteratorSchema().load(settings, **kwargs)
 
     @staticmethod
     def loads(settings, **kwargs):
-        """Create GraphTransliterator from JSON settings."""
+        """Create GraphTransliterator from JavaScript Object Notation (JSON) string.
+
+        Parameters
+        ----------
+        settings
+            JSON settings for GraphTransliterator
+
+        Returns
+        -------
+        GraphTransliterator
+            Graph Transliterator
+
+        Example
+        -------
+        >>> from graphtransliterator import GraphTransliterator
[…]
