### 2019-09-16 8a9b537
mbug fix and additional tests
