+++ requirements.txt
@@ -1,2 +1,2 @@
-marshmallow==3.0.5
+marshmallow==3.1.1
 pyyaml==5.1.2
