### 2019-09-17 8c40f9e
updates to paper
