+++ paper/paper.bib
@@ -7,6 +7,61 @@
   url = {http://userguide.icu-project.org/transforms/general}}
 }
 
+@manual{Barry,
+  title  = {ALA-LC Romanization Tables: Transliteration Schemes for Non-Roman Scripts},
+  author = {{Barry}, R. K.},
+  organization = {Library of Congress & American Library Association},
+  url    = "https://www.loc.gov/catdir/cpso/roman.html",
+  year   = "1997"
+}
+
+@article{prabhakar2018machine,
+  title={Machine transliteration and transliterated text retrieval: a survey},
+  author={Prabhakar, Dinesh Kumar and Pal, Sukomal},
+  journal={S{\=a}dhan{\=a}},
+  volume={43},
+  number={6},
+  pages={93},
+  year={2018},
+  publisher={Springer}
+}
+@inproceedings{durrani2010hindi,
+  title={Hindi-to-Urdu machine translation through transliteration},
+  author={Durrani, Nadir and Sajjad, Hassan and Fraser, Alexander and Schmid, Helmut},
+  booktitle={Proceedings of the 48th Annual meeting of the Association for Computational Linguistics},
+  pages={465--474},
+  year={2010},
+  organization={Association for Computational Linguistics}
+}
+
+@inproceedings{saini2008shahmukhi,
+  title={Shahmukhi to Gurmukhi transliteration system},
+  author={Saini, Tejinder Singh and Lehal, Gurpreet Singh and Kalra, Virinder S},
+  booktitle={22nd International Conference on on Computational Linguistics: Demonstration Papers},
+  pages={177--180},
+  year={2008},
+  organization={Association for Computational Linguistics}
+}
+@inproceedings{merhav-ash-2018-design,
+    title = "Design Challenges in Named Entity Transliteration",
+    author = "Merhav, Yuval  and
+      Ash, Stephen",
+    booktitle = "Proceedings of the 27th International Conference on Computational Linguistics",
+    month = aug,
+    year = "2018",
+    address = "Santa Fe, New Mexico, USA",
+    publisher = "Association for Computational Linguistics",
+    url = "https://www.aclweb.org/anthology/C18-1053",
+    pages = "630--640",
+    abstract = "We analyze some of the fundamental design challenges that impact the development of a multilingual state-of-the-art named entity transliteration system, including curating bilingual named entity datasets and evaluation of multiple transliteration methods. We empirically evaluate the transliteration task using the traditional weighted finite state transducer (WFST) approach against two neural approaches: the encoder-decoder recurrent neural network method and the recent, non-sequential Transformer method. In order to improve availability of bilingual named entity transliteration datasets, we release personal name bilingual dictionaries mined from Wikidata for English to Russian, Hebrew, Arabic, and Japanese Katakana. Our code and dictionaries are publicly available.",
+}
+@inproceedings{chen2018report,
+  title={Report of NEWS 2018 Named Entity Transliteration Shared Task},
+  author={Chen, Nancy and Banchs, Rafael E and Zhang, Min and Duan, Xiangyu and Li, Haizhou},
+  booktitle={Proceedings of the Seventh Named Entities Workshop},
+  pages={55--73},
+  year={2018}
+}
 
 @Manual{YAMLAintMarkup,
   title = {{{YAML~Ain}}'t~{{Markup~Language}}~({{YAML}}\texttrademark{}) {{Version}}~1.2},
