### 2019-09-17 b0a6927
updates to paper
