+++ paper/paper.bib
@@ -8,7 +8,7 @@
 }
 
 @manual{Barry,
-  title  = {ALA-LC Romanization Tables: Transliteration Schemes for Non-Roman Scripts},
+  title  = {ALA-{LC} Romanization Tables: Transliteration Schemes for Non-Roman Scripts},
   author = {{Barry}, R. K.},
   organization = {Library of Congress & American Library Association},
   url    = "https://www.loc.gov/catdir/cpso/roman.html",
