### 2019-09-17 #1717 @seanpue
Thanks @vc1492a for your generous and helpful review!

I have made the following changes to the oode and documentation:
* Fixed installation issues (using your pull request)
* Added list of dependencies to docs/installation.rst under subsection "Required Modules"
* Added badge for supported Python versions to README.rst

I have also added references and made changes to the first paragraph of the paper following your suggestion to better elaborate the use cases of transliteration.

I hope that resolves the issues. Thanks so much. (P.S. Would love to hear later about the potential usages!)
