+++ docs/index.rst
@@ -1,11 +1,9 @@
-Graph Transliterator Documentation
-==================================
-
+.. include:: ../README.rst
 
 .. toctree::
    :maxdepth: 2
 
-   readme
+
    installation
    usage
    tutorial
