+++ README.rst
@@ -61,8 +61,35 @@ Features
 * Permits **pruning of rules** with certain productions
 * **Validates**, as well as **serializes** to and **deserializes** from JSON
   and Python data types, using accessible
-  `marshmallow <https://marshmallow.readthedocs.io/en/stable/>`_ schemas
+  `marshmallow <https://marshmallow.readthedocs.io/>`_ schemas
 * Provides **full support for Unicode**, including Unicode **character names**
   in the "easy reading" YAML format
 * Constructs and uses a **directed tree** and performs a **best-first search**
   to find the most specific transliteration rule in a given context
+
+Sample Code and Graph
+---------------------
+>>> from graphtransliterator import GraphTransliterator
+>>> GraphTransliterator.from_yaml("""
+...     tokens:
+...       h: [consonant]
+...       i: [vowel]
+...       " ": [whitespace]
+...     rules:
+...       h: \N{LATIN SMALL LETTER TURNED I}
+...       i: \N{LATIN SMALL LETTER TURNED H}
+...       <whitespace> i: \N{LATIN CAPITAL LETTER TURNED h}
+...       (<whitespace> h) i: \N{LATIN SMALL LETTER TURNED H}!
+...     onmatch_rules:
+...       - <whitespace> + <consonant>: "¡"
+...     whitespace:
+...       default: " "
+...       consolidate: true
+...       token_class: whitespace
+...     metadata:
+...       title: "Upside Down Greeting Transliterator"
+...       version: "1.0"
+... """).transliterate("hi")
+'¡ᴉɥ!'
+
+.. image:: https://raw.githubusercontent.com/seanpue/graphtransliterator/master/docs/_static/sample_graph.png
