### 2019-09-17 79ab840
Bump version: 0.3.6 → 0.3.7
