### 2019-09-19 #1717 @rlskoeser
I have reviewed the software, documentation, and paper. The only checklist item I'm not able to sign off on is a statement of need included with the project documentation. As I understand it, a statement of need should be included in the readme or project documentation, but the only place I could find this was in the article. I suggest adapting the summary information from the beginning of the article.

Other things I noted:
- I was surprised not to find installation instructions included in the readme (at least the most common version); would be better if people didn't have to click through to read the docs to find that. Based on the [reviewing guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html) it sounds like this should be included in the readme, but I checked it off since installation is well-documented in the read the docs documentation.
- The detailed tutorial is well-written and very helpful, but it's difficult to cut and paste examples from the python console because of the `...` included on the beginning of continuing lines. I tried to generate a file I could load using the `from_yaml_file` method based on the annotated sections you provide in a tutorial. First I got an error because I included the short set of rules, which looks like it has a typo (PR opened). Now that I've corrected that, I'm getting a validation error that I'm not sure how to troubleshoot. (I can't tell if this is a discrepancy between the versions in the tutorial or something I've done wrong.)
```
marshmallow.exceptions.ValidationError: {'rules': ['Invalid token "," in tokens of rule TransliterationRule(production=\',\', prev_classes=None, prev_tokens=None, tokens=[\',\'], next_tokens=None, next_classes=None, cost=0.5849625007211562)']}
```
Perhaps you could include the complete yaml file as a link that could be downloaded for use with the tutorial, and then people could compare your working version with the annotated samples in the tutorial? That would also allow you to omit the full lists of tokens rules from the tutorial, which might make it easier to follow.
- In your contributing doc, the link on GitHub is broken but it works fine on readthedocs. I'm not sure if there's a way to make it work in both places
- The directions for contributing say to use the Black code formatter, but it isn't included in the dev requirements. You have a note about pip installing them, but it seems like it would be better to add black to the dev requirements and update the contributing instructions to install development dependencies from that file.
