### 2019-09-20 a8c37b3
minor fixes
