+++ paper/paper.bib
@@ -25,6 +25,7 @@
   year={2018},
   publisher={Springer}
 }
+
 @inproceedings{durrani2010hindi,
   title={Hindi-to-Urdu machine translation through transliteration},
   author={Durrani, Nadir and Sajjad, Hassan and Fraser, Alexander and Schmid, Helmut},
@@ -42,6 +43,7 @@
   year={2008},
   organization={Association for Computational Linguistics}
 }
+
 @inproceedings{merhav-ash-2018-design,
     title = "Design Challenges in Named Entity Transliteration",
     author = "Merhav, Yuval  and
@@ -55,6 +57,7 @@
     pages = "630--640",
     abstract = "We analyze some of the fundamental design challenges that impact the development of a multilingual state-of-the-art named entity transliteration system, including curating bilingual named entity datasets and evaluation of multiple transliteration methods. We empirically evaluate the transliteration task using the traditional weighted finite state transducer (WFST) approach against two neural approaches: the encoder-decoder recurrent neural network method and the recent, non-sequential Transformer method. In order to improve availability of bilingual named entity transliteration datasets, we release personal name bilingual dictionaries mined from Wikidata for English to Russian, Hebrew, Arabic, and Japanese Katakana. Our code and dictionaries are publicly available.",
 }
+
 @inproceedings{chen2018report,
   title={Report of NEWS 2018 Named Entity Transliteration Shared Task},
   author={Chen, Nancy and Banchs, Rafael E and Zhang, Min and Duan, Xiangyu and Li, Haizhou},
@@ -62,68 +65,9 @@
   pages={55--73},
   year={2018}
 }
-
 @Manual{YAMLAintMarkup,
   title = {{{YAML~Ain}}'t~{{Markup~Language}}~({{YAML}}\texttrademark{}) {{Version}}~1.2},
   author = {{Ben-Kiki}, O. and {Evans}, C. and {döt Net}, I.},
   year = {2009},
   url = {https://yaml.org/spec/1.2/spec.html}
-}
-@misc{YAMLAintMarkup,
-
-  howpublished = {https://yaml.org/spec/1.2/spec.html},
-  file = {/Users/seanpue/Zotero/storage/6FV6WL7Y/spec.html}
-}
-
-@article{Pearson:2017,
-  	Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
-  	Adsurl = {http://adsabs.harvard.edu/abs/2017arXiv170304627P},
-  	Archiveprefix = {arXiv},
-  	Author = {{Pearson}, S. and {Price-Whelan}, A.~M. and {Johnston}, K.~V.},
-  	Eprint = {1703.04627},
-  	Journal = {ArXiv e-prints},
-  	Keywords = {Astrophysics - Astrophysics of Galaxies},
-  	Month = mar,
-  	Title = {{Gaps in Globular Cluster Streams: Pal 5 and the Galactic Bar}},
-  	Year = 2017
-}
-
-@book{Binney:2008,
-  	Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
-  	Adsurl = {http://adsabs.harvard.edu/abs/2008gady.book.....B},
-  	Author = {{Binney}, J. and {Tremaine}, S.},
-  	Booktitle = {Galactic Dynamics: Second Edition, by James Binney and Scott Tremaine.~ISBN 978-0-691-13026-2 (HB).~Published by Princeton University Press, Princeton, NJ USA, 2008.},
-  	Publisher = {Princeton University Press},
-  	Title = {{Galactic Dynamics: Second Edition}},
-  	Year = 2008
-}
-
-@article{gaia,
-    author = {{Gaia Collaboration}},
[…]
