+++ graphtransliterator/transliterators/sample/__init__.py
@@ -1,0 +1,3 @@
+class Sample:
+    def init():
+        return __file__
