+++ graphtransliterator/transliterators/README.rst
@@ -1,0 +1,48 @@
+==============================
+Bundled Transliterators--DRAFT
+==============================
+This directory contains bundled transliterators. The root directory contains the
+transliterators as YAML files. The subdirectory `tests` contains tests for each
+transliterator.
+
+If you are creating a transliterator, see the `sample` subdirectory.
+
+# A test is required, and it must have complete coverage of the
+# graph, meaning every node must be visited over the course of the tests.
+
+Naming Conventions
+------------------
+Easy-reading YAML file, if used: `./SOURCE_TO_TARGET/SOURCE_to_TARGET.yaml`
+# JSON file: `./SOURCE_TO_TARGET/SOURCE_to_TARGET.json` - generated by script?
+# Initialization file: `./SOURCE_TO_TARGET/__init__.py`?
+Test file: `./SOURCE_TO_TARGET/tests/SOURCE_TO_TARGET.py`
+# Long Description file: SOURCE_TO_TARGET/SOURCE_TO_TARGET.rst
+
+Metadata Requirements
+---------------------
+Each transliterator can include the following metadata fields. These fields are a
+subset of `setuptools`. Long descriptions are not currently included.
+
+name (`str`)
+  Name of the transliterator. It must match the name of the YAML file. If possible,
+  it should be in lowercase and follow the format: `SOURCE_to_TARGET.yaml`, where
+  `SOURCE` is the original script/language/etc. and `TARGET` is the destination
+  script/language/etc.
+version	(`str`, optional)
+  Version of the transliterator. Semantic versioning (https://semver.org) is
+  strongly recommended.
+url	(`str`, optional)
+  URL for the transliterator, e.g. Git. repository.
+author (`str`, optional)
+  Author of the transliterator
+author_email (`str`, optional)
+  E-mail address of the author.
+maintainer (`str`, optional)
+  Name of the maintainer.
+maintainer_email (`str`, optional)
+  E-mail address of the maintainer.
+license (`str`, optional)
+  License of the transliterator. An open-source license is required for inclusion in
+  this project.
+keywords (`list` of `str`, optional)
+  List of keywords
