### 2019-09-29 3033dba
working with docs in process with jupyter-sphinx
