+++ docs/bundled.rst
@@ -1,0 +1,1 @@
+.. include:: ../graphtransliterator/transliterators/README.rst

[…]
