### 2019-09-30 0ff03ce
history.rst fixes
