### 2019-09-30 #1717 @seanpue
I'm close to finishing these corrections. It will probably take at least another week.

I have added  bundled transliterators requiring tests that cover all nodes an edges of graphs, as well as OnMatchRules, if applicable. I will have to update the essay, as well.

I will also be cleaning up the documentation and other points following the excellent recommendations of @rlskoeser.

This review process has been so helpful! Thanks, @vc1492a  and @rlskoeser!
