### 2019-09-30 #1717 @gkthiruvathukal
@seanpue: Thanks for your follow up and willingness to address all review feedback.

@vc1492a and @rlskoeser: Thanks for your thorough reviews. This feels like a model of how reviews should be done. Even though the work is clearly well-done and polished, there were minor issues raised to ensure the submission leads to a great outcome.

I feel optimistic this is heading toward acceptance and look forward to recommending same when @seanpue finishes making the minor revisions.  Separately, I rarely jump in with any feedback before the reviewers do, but I felt pretty optimistic about this submission when I first saw it.  It's nice to see this research area represented within JOSS.
