### 2019-09-30 d40f958
adjusting path in docs/conf.py
