### 2019-09-30 59833ee
fix to docs/cofnf.py
