+++ docs/conf.py
@@ -18,14 +18,11 @@
 # relative to the documentation root, use os.path.abspath to make it
 # absolute, like shown here.
 #
-impoprt sys
-sys.path.insert(0, os.path.abspath('.'))
-sys.path.insert(0, os.path.abspath('../'))
-import graphtransliterator
 import os
 import sys
 
 sys.path.insert(0, os.path.abspath(".."))
+import graphtransliterator  # noqa
 
 # -- General configuration ---------------------------------------------
 
