### 2019-09-30 82d8072
added more paths to docs/conf.py for jupyter-sphinx
