+++ docs/conf.py
@@ -22,6 +22,8 @@ import os
 import sys
 
 sys.path.insert(0, os.path.abspath(".."))
+package_path = os.path.abspath("../..")
+os.environ["PYTHONPATH"] = ":".join((package_path, os.environ.get("PYTHONPATH", "")))
 import graphtransliterator  # noqa
 
 # -- General configuration ---------------------------------------------
