+++ docs/conf.py
@@ -22,7 +22,8 @@ import os
 import sys
 
 sys.path.insert(0, os.path.abspath(".."))
-sys.path.insert(0, os.path.abspath("../.."))  # /../docs/
+sys.path.insert(0, os.path.abspath("../.."))
+sys.path.insert(0, os.path.abspath("../../.."))  # /../docs/
 # package_path = os.path.abspath("../..")
 # os.environ["PYTHONPATH"] = ":".join((package_path, os.environ.get("PYTHONPATH", "")))
 import graphtransliterator  # noqa
