+++ docs/conf.py
@@ -21,7 +21,9 @@
 import os
 import sys
 
-sys.path.insert(0, os.path.abspath("../.."))
+sys.path.insert(
+    0, os.path.abspath("..")
+)  # necessary to import graphtransilterator here
 
 # package_path = os.path.abspath("../..")
 # os.environ["PYTHONPATH"] = ":".join((package_path, os.environ.get("PYTHONPATH", "")))
