### 2019-10-01 a5a4fb2
added setuptools to requirements.txt
