### 2019-10-01 92ba263
Pin setuptools to latest version 41.2.0
