+++ requirements.txt
@@ -1,3 +1,3 @@
 marshmallow==3.2.0
 pyyaml==5.1.2
-setuptools
+setuptools==41.2.0
