### 2019-10-01 dee8dfd
trying __version__ instead of pkgresource
