### 2019-10-02 98f45e7
updates; some errors with transition to jupyter-sphinx
