+++ graphtransliterator/core.py
@@ -159,11 +159,14 @@ class GraphTransliterator:
 
     Example
     -------
-    >>> from graphtransliterator import *
-    >>> settings = {'tokens': {'a': {'vowel'}, ' ': {'wb'}}, 'onmatch_rules': [OnMatchRule(prev_classes=['vowel'], next_classes=['vowel'], production=',')], 'rules': [TransliterationRule(production='A', prev_classes=None, prev_tokens=None, tokens=['a'], next_tokens=None, next_classes=None, cost=0.5849625007211562), TransliterationRule(production=' ', prev_classes=None, prev_tokens=None, tokens=[' '], next_tokens=None, next_classes=None, cost=0.5849625007211562)], 'metadata': {'author': 'Author McAuthorson'}, 'whitespace': WhitespaceRules(default=' ', token_class='wb', consolidate=False)}
-    >>> gt = GraphTransliterator(**settings)
-    >>> gt.transliterate('a')
-    'A'
+    .. jupyter-execute::
+
+
+      from graphtransliterator import GraphTransliterator, OnMatchRule, TransliterationRule, WhitespaceRules
+      settings = {'tokens': {'a': {'vowel'}, ' ': {'wb'}}, 'onmatch_rules': [OnMatchRule(prev_classes=['vowel'], next_classes=['vowel'], production=',')], 'rules': [TransliterationRule(production='A', prev_classes=None, prev_tokens=None, tokens=['a'], next_tokens=None, next_classes=None, cost=0.5849625007211562), TransliterationRule(production=' ', prev_classes=None, prev_tokens=None, tokens=[' '], next_tokens=None, next_classes=None, cost=0.5849625007211562)], 'metadata': {'author': 'Author McAuthorson'}, 'whitespace': WhitespaceRules(default=' ', token_class='wb', consolidate=False)}
+      gt = GraphTransliterator(**settings)
+      gt.transliterate('a')
+
 
     See Also
     --------
@@ -353,33 +356,40 @@ class GraphTransliterator:
         Examples
         --------
 
-        >>> from graphtransliterator import GraphTransliterator
-        >>> gt = GraphTransliterator.from_yaml('''
-        ...         tokens:
-        ...             a: []
-        ...             a a: []
-        ...             ' ': [wb]
-        ...         rules:
-        ...             a: <A>
-        ...             a a: <AA>
-        ...         whitespace:
-        ...             default: ' '
-        ...             consolidate: True
-        ...             token_class: wb
-        ... ''')
-        >>> tokens = gt.tokenize("aa")
-        >>> tokens # whitespace added to ends
-        [' ', 'a', 'a', ' ']
-        >>> gt.match_at(1, tokens) # returns index to rule
-        0
-        >>> gt.rules[gt.match_at(1, tokens)] # actual rule
-        TransliterationRule(production='<AA>', prev_classes=None, prev_tokens=None, tokens=['a', 'a'], next_tokens=None, next_classes=None, cost=0.41503749927884376)
-        >>> gt.match_at(1, tokens, match_all=True) # index to rules, with match_all
-        [0, 1]
-        >>>
-        >>> [gt.rules[_] for _ in gt.match_at(1, tokens, match_all=True)]
-        [TransliterationRule(production='<AA>', prev_classes=None, prev_tokens=None, tokens=['a', 'a'], next_tokens=None, next_classes=None, cost=0.41503749927884376), TransliterationRule(production='<A>', prev_classes=None, prev_tokens=None, tokens=['a'], next_tokens=None, next_classes=None, cost=0.5849625007211562)]
-        >>>
+        .. jupyter-execute::
+
+          gt = GraphTransliterator.from_yaml('''
+                  tokens:
+                      a: []
+                      a a: []
[…]
