### 2019-10-02 dd86e31
update readme.rst
probablem with code-block or with linenos
