+++ docs/usage.rst
@@ -41,6 +41,7 @@ Here is a quick sample that parameterizes :class:`GraphTransliterator` using an 
 reading YAML string (with comments):
 
 .. jupyter-execute::
+  :linenos:
 
   from graphtransliterator import GraphTransliterator
   yaml_ = """
@@ -62,10 +63,12 @@ reading YAML string (with comments):
   gt_one.transliterate('a')
 
 .. jupyter-execute::
+  :continue_linenos:
 
   gt_one.transliterate('bb')
 
 .. jupyter-execute::
+  :continue_linenos:
 
   gt_one.transliterate('aabb')
 
@@ -82,6 +85,7 @@ into a list of tokens. This is done internally using
 :meth:`GraphTransliterator.tokenize`:
 
   .. jupyter-execute::
+    :continue_linenos:
 
     gt_one.tokenize('abba')
 
@@ -92,6 +96,7 @@ tokens.
 Tokens can be more than one character, and longer tokens are matched first:
 
 .. jupyter-execute::
+  :linenos:
 
   yaml_ = """
     tokens:
@@ -110,10 +115,12 @@ Tokens can be more than one character, and longer tokens are matched first:
   gt_two.transliterate('a')
 
 .. jupyter-execute::
+  :continue_linenos:
 
   gt_two.transliterate('aa')
 
 .. jupyter-execute::
+  :continue_linenos:
 
   gt_two.transliterate('aaa')
 
@@ -128,6 +135,7 @@ to a set of token classes. They can be accessed using
 :attr:`GraphTransliterator.tokens`:
 
 .. jupyter-execute::
+  :continue_linenos:
 
   gt_two.tokens
 
@@ -226,6 +234,7 @@ decreases depending on the number of tokens required by the rule. More tokens de
 the cost of a rule causing it to be matched first:
 
 .. jupyter-execute::
+  :linenos:
 
   yaml_ = """
     tokens:
@@ -249,6 +258,7 @@ the cost of a rule causing it to be matched first:
   gt_three.transliterate("ab")  # should match rule "a b"
 
 .. jupyter-execute::
+  :continue_linenos:
 
   gt_three.transliterate("cab") # should match rules: "c", and "<class_of_c> a b"
 
@@ -258,6 +268,7 @@ These can be accessed using :attr:`GraphTransliterator.rules`. Rules are sorted 
 lowest to highest:
 
 .. jupyter-execute::
+  :continue_linenos:
 
   gt_three.rules
 
@@ -283,6 +294,7 @@ replaces any sequential whitespace tokens in the input string with the default
 whitespace character. At the start and end of input, it removes any whitespace:
 
 .. jupyter-execute::
+  :linenos:
 
   yaml_ = """
     tokens:
@@ -303,14 +315,17 @@ whitespace character. At the start and end of input, it removes any whitespace:
   gt.transliterate('a')   # whitespace present at start of string
 
 .. jupyter-execute::
+  :continue_linenos:
 
   gt.transliterate('aa')  # whitespace present at start and end of string
 
 .. jupyter-execute::
+  :continue_linenos:
 
   gt.transliterate(' a')  # consolidate removes whitespace at start of string
 
 .. jupyter-execute::
+  :continue_linenos:
 
   gt.transliterate('a ')  # consolidate removes whitespace at end of string
 
@@ -319,6 +334,7 @@ Whitespace settings are stored internally as :class:`WhitespaceRules` and can be
 accessed using :attr:`GraphTransliterator.whitespace`:
 
 .. jupyter-execute::
+  :continue_linenos:
 
   gt.whitespace
 
[…]
