### 2019-10-04 49bce8b
updates to css for jupyter-sphinx and some doc updates
