+++ docs/usage.rst
@@ -63,12 +63,12 @@ reading YAML string (with comments):
   gt_one.transliterate('a')
 
 .. jupyter-execute::
-  :continue_linenos:
+  :continue-linenos:
 
   gt_one.transliterate('bb')
 
 .. jupyter-execute::
-  :continue_linenos:
+  :continue-linenos:
 
   gt_one.transliterate('aabb')
 
@@ -84,10 +84,10 @@ During transliteration, Graph Transliterator first attempts to convert the input
 into a list of tokens. This is done internally using
 :meth:`GraphTransliterator.tokenize`:
 
-  .. jupyter-execute::
-    :continue_linenos:
+.. jupyter-execute::
+  :continue-linenos:
 
-    gt_one.tokenize('abba')
+  gt_one.tokenize('abba')
 
 
 Note that the default whitespace  token is added to the start and end of the input
@@ -115,12 +115,12 @@ Tokens can be more than one character, and longer tokens are matched first:
   gt_two.transliterate('a')
 
 .. jupyter-execute::
-  :continue_linenos:
+  :continue-linenos:
 
   gt_two.transliterate('aa')
 
 .. jupyter-execute::
-  :continue_linenos:
+  :continue-linenos:
 
   gt_two.transliterate('aaa')
 
@@ -135,7 +135,7 @@ to a set of token classes. They can be accessed using
 :attr:`GraphTransliterator.tokens`:
 
 .. jupyter-execute::
-  :continue_linenos:
+  :continue-linenos:
 
   gt_two.tokens
 
@@ -258,7 +258,7 @@ the cost of a rule causing it to be matched first:
   gt_three.transliterate("ab")  # should match rule "a b"
 
 .. jupyter-execute::
-  :continue_linenos:
+  :continue-linenos:
 
   gt_three.transliterate("cab") # should match rules: "c", and "<class_of_c> a b"
 
@@ -268,7 +268,7 @@ These can be accessed using :attr:`GraphTransliterator.rules`. Rules are sorted 
 lowest to highest:
 
 .. jupyter-execute::
-  :continue_linenos:
+  :continue-linenos:
 
   gt_three.rules
 
@@ -315,17 +315,17 @@ whitespace character. At the start and end of input, it removes any whitespace:
   gt.transliterate('a')   # whitespace present at start of string
 
 .. jupyter-execute::
-  :continue_linenos:
+  :continue-linenos:
 
   gt.transliterate('aa')  # whitespace present at start and end of string
 
 .. jupyter-execute::
-  :continue_linenos:
+  :continue-linenos:
 
   gt.transliterate(' a')  # consolidate removes whitespace at start of string
 
 .. jupyter-execute::
-  :continue_linenos:
+  :continue-linenos:
 
   gt.transliterate('a ')  # consolidate removes whitespace at end of string
 
@@ -334,7 +334,7 @@ Whitespace settings are stored internally as :class:`WhitespaceRules` and can be
 accessed using :attr:`GraphTransliterator.whitespace`:
 
 .. jupyter-execute::
-  :continue_linenos:
+  :continue-linenos:
 
   gt.whitespace
 
@@ -379,7 +379,7 @@ On Match rules are stored internally as a :class:`OnMatchRule` and can be access
 :attr:`GraphTransliterator.onmatch_rules`:
 
 .. jupyter-execute::
-  :continue_linenos:
+  :continue-linenos:
 
   gt.onmatch_rules
 
@@ -440,7 +440,7 @@ the hexadecimal character code):
   gt.transliterate('b')
 
 .. jupyter-execute::
-  :continue_linenos:
+  :continue-linenos:
 
   gt.transliterate('c')
 
[…]
