+++ graphtransliterator/transliterators/README.rst
@@ -2,161 +2,10 @@
 Bundled Transliterators
 =======================
 
-Graph Transliterator includes bundled transliterators in a :class:`Bundled` subclass of
-:class:`GraphTransliterator` that can be used as follows:
+Bundled transliterators are found here, in subdirectories of
+``graphtransliterator/transliterators``.
 
-.. jupyter-execute::
-  :linenos:
+See the documentation for information on how to use and create them, or take a look in
+``graphtransliterator/transliterators/example``.
 
-  import graphtransliterator.transliterators as transliterators
-  example_transliterator = transliterators.Example()
-  example_transliterator.transliterate('a')
-
-To access transliterator classes, use the iterator
-:func:`transliterators.iter_transliterators`:
-
-.. jupyter-execute::
-  :continue_linenos:
-
-  bundled_iterator = transliterators.iter_transliterators()
-  next(bundled_iterator)
-
-To access the names of transliterator classes, use the iterator
-:func:`transliterators.iter_transliterators`:
-
-.. jupyter-execute::
-  :continue_linenos:
-
-  bundled_names_iterator = transliterators.iter_names()
-  next(bundled_names_iterator)
-
-The actual bundled transliterators are submodules of
-:class:`graphtransliterator.transliterators`, but they are loaded into the namespace
-of :class:`transliterators`:
-
-.. jupyter-execute::
-  :continue_linenos:
-
-  from graphtransliterator.transliterators import Example
-
-Each instance of :class:`Bundled` contains a :py:attr:`Bundled.directory` attribute:
-
-.. jupyter-execute::
-  :continue_linenos:
-
-  transliterator = Example()
-  transliterator.directory
-
-Each will contain an easy-reading YAML file that you can view, as well as a JSON
-dump of the transliterator for quick loading:
-
-.. jupyter-execute::
-  :hide-code:
-
-  import os
-  with open(os.path.join(transliterator.directory, "example.yaml"), "r") as f:
-    print("--Easy-reading YAML (for clarity, development, and debugging)--")
-    print(f.read()+"\n")
-  with open(os.path.join(transliterator.directory, "example.json"), "r") as f:
-    print("--JSON (for speed)--")
-    print(f.read())
-
-Test Coverage of Bundled Transliterators
-----------------------------------------
-
-Each bundled transliterators requires rigorous testing: every node and edge, as
-well as any onmatch rules, if applicable, must be visited. A separate subclass,
-:class:`CoverageTransliterator` (in `core.py`) of :class:`GraphTransliterator` is used
-during testing (see `tests/test_coverage.py`).
-It logs visits to nodes, edges, and onmatch rules. The tests are found in a subdirectory
-of the transliterator called "tests" and consists of a YAML file consisting of a
-dictionary keyed from transliteration input to correct output, e.g.:
-
-.. jupyter-execute::
-  :hide-code:
-
-  with open(os.path.join(transliterator.directory, "tests/example_tests.yaml"), "r") as f:
-    print("-- YAML Tests for Example (with complete graph and onmatch rule coverage) --\n")
-    print(f.read())
-
-Once the checks are completed, the tests check that all components of the graph and all
-of the onmatch rules have been visited.
-
-Class Structure and Naming Conventions
---------------------------------------
-Each transliterator must include a class definition in a submodule of
-:class:`transliterators`.
-
-The class name of each transliterator must be unique and follow camel-case conventions,
-e.g. `SourceToTarget`. File and directory names should, if applicable, be lowercased as
-`source_to_target`.
-
-The bundled files should follow this directory structure, where {{source_to_target}} is
-the name of the transliterator:
-
-
-.. code::
-
[…]
