+++ docs/index.rst
@@ -1,3 +1,13 @@
+.. -------------------------------------------------------------------------------------
+.. Note:
+..     This is a documentation source file for Graph Transliterator.
+..     Certain links and other features will not be accessible from here.
+.. Links:
+..     - Documentation: https://graphtransliterator.readthedocs.org
+..     - PyPI: https://pypi.org/project/graphtransliterator/
+..     - Repository: https://github.com/seanpue/graphtransliterator/
+.. -------------------------------------------------------------------------------------
+
 .. include:: ../README.rst
 
 .. toctree::
