+++ docs/conf.py
@@ -106,7 +106,7 @@ todo_include_todos = False
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = "sphinx_rtd_theme"
+html_theme = "alabaster"#sphinx_rtd_theme"
 
 # Theme options are theme-specific and customize the look and feel of a
 # theme further.  For a list of options available for each theme, see the
@@ -189,8 +189,11 @@ texinfo_documents = [
     )
 ]
 
-# testing autosummary
-# autosummary_generate = True
+# jupyter-sphinx SettingsSchema
 
+jupyter_sphinx_linenos = True
+jupyter_sphinx_continue_linenos = True
+#jupyter_sphinx_truncate_traceback = True
+# jupyter_sphinx_stderr_classes = ["admonition", "warning"]
 def setup(app):
    app.add_stylesheet('custom.css')
