+++ HISTORY.rst
@@ -19,6 +19,9 @@ History
 
 #.#.# (####-##-##)
 ------------------
+
+* added preface to documentation source files with links to production version, etc.
+  Thanks :user:`rlskoeser`.
 * added custom css for jupyter-sphinx cells
 * added jupyter-sphinx documentation with line numbering
 * removed pkg_resources as source for version due to problem with loading from
