+++ docs/usage.rst
@@ -73,12 +73,10 @@ reading YAML string (with comments):
   gt_one.transliterate('a')
 
 .. jupyter-execute::
-  :continue-linenos:
 
   gt_one.transliterate('bb')
 
 .. jupyter-execute::
-  :continue-linenos:
 
   gt_one.transliterate('aabb')
 
@@ -95,7 +93,6 @@ into a list of tokens. This is done internally using
 :meth:`GraphTransliterator.tokenize`:
 
 .. jupyter-execute::
-  :continue-linenos:
 
   gt_one.tokenize('abba')
 
@@ -125,12 +122,10 @@ Tokens can be more than one character, and longer tokens are matched first:
   gt_two.transliterate('a')
 
 .. jupyter-execute::
-  :continue-linenos:
 
   gt_two.transliterate('aa')
 
 .. jupyter-execute::
-  :continue-linenos:
 
   gt_two.transliterate('aaa')
 
@@ -145,7 +140,6 @@ to a set of token classes. They can be accessed using
 :attr:`GraphTransliterator.tokens`:
 
 .. jupyter-execute::
-  :continue-linenos:
 
   gt_two.tokens
 
@@ -268,7 +262,6 @@ the cost of a rule causing it to be matched first:
   gt_three.transliterate("ab")  # should match rule "a b"
 
 .. jupyter-execute::
-  :continue-linenos:
 
   gt_three.transliterate("cab") # should match rules: "c", and "<class_of_c> a b"
 
@@ -278,7 +271,6 @@ These can be accessed using :attr:`GraphTransliterator.rules`. Rules are sorted 
 lowest to highest:
 
 .. jupyter-execute::
-  :continue-linenos:
 
   gt_three.rules
 
@@ -325,17 +317,14 @@ whitespace character. At the start and end of input, it removes any whitespace:
   gt.transliterate('a')   # whitespace present at start of string
 
 .. jupyter-execute::
-  :continue-linenos:
 
   gt.transliterate('aa')  # whitespace present at start and end of string
 
 .. jupyter-execute::
-  :continue-linenos:
 
   gt.transliterate(' a')  # consolidate removes whitespace at start of string
 
 .. jupyter-execute::
-  :continue-linenos:
 
   gt.transliterate('a ')  # consolidate removes whitespace at end of string
 
@@ -344,7 +333,6 @@ Whitespace settings are stored internally as :class:`WhitespaceRules` and can be
 accessed using :attr:`GraphTransliterator.whitespace`:
 
 .. jupyter-execute::
-  :continue-linenos:
 
   gt.whitespace
 
@@ -389,7 +377,6 @@ On Match rules are stored internally as a :class:`OnMatchRule` and can be access
 :attr:`GraphTransliterator.onmatch_rules`:
 
 .. jupyter-execute::
-  :continue-linenos:
 
   gt.onmatch_rules
 
@@ -450,7 +437,6 @@ the hexadecimal character code):
   gt.transliterate('b')
 
 .. jupyter-execute::
-  :continue-linenos:
 
   gt.transliterate('c')
 
@@ -495,6 +481,10 @@ Graph Transliterator, by default, will check for ambiguity in its transliteratio
 If two rules of the same cost would match the same string(s) and those strings would not
 be matched by a less costly rule, an :exc:`AmbiguousTransliterationRulesException`
 occurs. Details of all exceptions will be reported as a :meth:`logging.warning`:
+
+.. jupyter-execute::
+
+  %xmode Minimal
 
 .. jupyter-execute::
   :raises: AmbiguousTransliterationRulesException
@@ -631,7 +621,6 @@ the token index will be advanced. Otherwise, there will be a
[…]
