### 2019-10-14 4b5ef9e
restored rtd theme to docs
