+++ docs/conf.py
@@ -195,3 +195,4 @@ jupyter_sphinx_linenos = True
 jupyter_sphinx_continue_linenos = True
 def setup(app):
    app.add_stylesheet('custom.css')
+   app.add_stylesheet('jupyter-sphinx.css')
