### 2019-10-15 58679ff
new jupyter-sphinx branch
