### 2019-10-15 e56ed0f
hid xmode code/output
