### 2019-10-15 8456462
updates to usage.rst
