+++ docs/conf.py
@@ -191,8 +191,9 @@ texinfo_documents = [
 
 # jupyter-sphinx SettingsSchema
 
-jupyter_sphinx_linenos = True
+# jupyter_sphinx_linenos = True
 jupyter_sphinx_continue_linenos = True
-def setup(app):
-   app.add_stylesheet('custom.css')
-   app.add_stylesheet('jupyter-sphinx.css')
+##def setup(app):
+##   app.add_stylesheet('custom.css')
+##   app.add_stylesheet('jupyter-sphinx.css')
+##
