### 2019-10-15 7af7142
updates to bundled.rst, may need to change file permissions
