### 2019-10-16 a39875d
updates to tutorial
