+++ docs/tutorial.rst
@@ -70,114 +70,20 @@ Graph Transliterator allows the use of Unicode character names in files using
 characters using that notation or directly. YAML will also unescape \\u####,
 where #### is the hexadecimal notation for a character.
 
+Here is a subsection of that definition:
 
 .. code-block:: yaml
 
   tokens:
     k: [consonant]
     kh: [consonant]
-    g: [consonant]
-    gh: [consonant]
-    ~N: [consonant]
     "\N{LATIN SMALL LETTER N WITH DOT ABOVE}": [consonant]
-    ch: [consonant]
-    chh: [consonant]
-    Ch: [consonant]
-    j: [consonant]
-    jh: [consonant]
-    ~n: [consonant]
-    T: [consonant]
-    Th: [consonant]
-    D: [consonant]
-    Dh: [consonant]
-    N: [consonant]
-    t: [consonant]
-    th: [consonant]
-    d: [consonant]
-    dh: [consonant]
-    n: [consonant]
-    ^n: [consonant]
-    p: [consonant]
-    ph: [consonant]
-    b: [consonant]
-    bh: [consonant]
-    m: [consonant]
-    y: [consonant]
-    r: [consonant]
-    R: [consonant]
-    l: [consonant]
-    ld: [consonant]
-    L: [consonant]
-    zh: [consonant]
-    v: [consonant]
-    sh: [consonant]
-    Sh: [consonant]
-    s: [consonant]
-    h: [consonant]
-    x: [consonant]
-    kSh: [consonant]
-    GY: [consonant]
-    j~n: [consonant]
-    dny: [consonant]
-    q: [consonant]
-    K: [consonant]
-    G: [consonant]
-    J: [consonant]
-    z: [consonant]
-    .D: [consonant]
-    .Dh: [consonant]
-    f: [consonant]
-    Y: [consonant]
     a: [vowel]
     aa: [vowel]
     A: [vowel]
-    i: [vowel]
-    ii: [vowel]
-    I: [vowel]
-    ee: [vowel]
-    u: [vowel]
-    uu: [vowel]
-    U: [vowel]
-    RRi: [vowel]
-    R^i: [vowel]
-    LLi: [vowel]
-    L^i: [vowel]
-    RRI: [vowel]
-    LLI: [vowel]
-    a.c: [vowel]
-    ^e: [vowel]
-    e: [vowel]
-    ai: [vowel]
-    A.c: [vowel]
-    ^o: [vowel]
-    o: [vowel]
-    au: [vowel]
     ' ': [wb,whitespace]
     "\t": [wb,whitespace]
-    .h: [wb]
-    H: [wb]
-    OM: [wb]
-    AUM: [wb]
-    '|': [wb]
-    '||': [wb]
-    '0': [wb]
-    '1': [wb]
-    '2': [wb]
-    '3': [wb]
-    '4': [wb]
-    '5': [wb]
-    '6': [wb]
-    '7': [wb]
-    '8': [wb]
-    '9': [wb]
-    Rs.: [wb]
-    ~Rs.: [wb]
-    .a: [wb]
-    a.e: [vowel]
[…]
