### 2019-10-22 #1717 @kthyng
Hi @seanpue how are your corrections coming along?
