### 2019-10-22 104981c
added bundled transliterators and cli
