+++ tests/test_coverage.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 
-# from graphtransliterator.coverage import CoverageTransliterator
 from graphtransliterator.exceptions import (
     IncompleteGraphCoverageException,
     IncompleteOnMatchRulesCoverageException,
@@ -11,20 +10,15 @@ import pytest
 
 def test_CoverageTransliterator():
     """Test coverage using transliterators.Example"""
-    # YAML:
-    # ------
     # tokens:
     #   a: [vowel]
     #   ' ': [whitespace]
+    #   b: [consonant]
     # rules:
     #   a: A
+    #   b: B
     #   ' ': ' '
-    # onmatch_rules:
-    #   - <vowel> + <vowel>: ","
-    # whitespace:
-    #   consolidate: False
-    #   default: " "
-    #   token_class: whitespace
+    #   (<consonant> a) b (a <consonant>):  "!B!"
     # -----
     covt = transliterators.Example(coverage=True)
     # unvisited should raise error
@@ -33,10 +27,13 @@ def test_CoverageTransliterator():
     # visiting only some nodes should raise error
     with pytest.raises(IncompleteGraphCoverageException):
         assert covt.transliterate("a") == "A"
+        assert covt.transliterate("b") == "B"
         covt.check_coverage()
     # visiting all nodes but no onmatch rules should raise error
     with pytest.raises(IncompleteOnMatchRulesCoverageException):
         assert covt.transliterate("a ") == "A "
+        assert covt.transliterate("b") == "B"
+        assert covt.transliterate("babab") == "BA!B!AB"
         covt.check_coverage()
     # visiting everything should not raise an error
     assert covt.transliterate("aa ") == "A,A "
@@ -44,8 +41,4 @@ def test_CoverageTransliterator():
     # clearing visited should raise an error when checking coverage
     with pytest.raises(IncompleteGraphCoverageException):
         covt.clear_visited()
-        covt.check_coverage()
-    # visiting all nodes but not using all onmatch rules should raise an error
-    with pytest.raises(IncompleteOnMatchRulesCoverageException):
-        assert covt.transliterate("a ") == "A "
         covt.check_coverage()
