+++ tests/test_cli.py
@@ -1,0 +1,128 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+"""Tests for `graphtransliterator` package."""
+import click
+from click.testing import CliRunner
+from graphtransliterator import cli
+from graphtransliterator import GraphTransliterator
+from io import StringIO
+import json
+import yaml
+
+
+def test_command_line_interface():
+    """Test the CLI."""
+    runner = CliRunner()
+    result = runner.invoke(cli.main)
+    assert result.exit_code == 0
+    assert "main" in result.output
+    help_result = runner.invoke(cli.main, ["--help"])
+    assert help_result.exit_code == 0
+    assert "--help  Show this message and exit." in help_result.output
+
+
+test_yaml = """
+    tokens:
+      a: [vowel]
+      ' ': [whitespace]
+    rules:
+      a: A
+      ' ': ' '
+    onmatch_rules:
+      - <vowel> + <vowel>: ","
+    whitespace:
+      consolidate: False
+      default: " "
+      token_class: whitespace
+"""
+
+
+def test_cli_transliterate_tests(tmpdir):
+    """Tests transliterate command, and loading of all formats."""
+    transliterator = GraphTransliterator.from_yaml(test_yaml)
+
+    runner = CliRunner()
+
+    # test bundled
+    bundled_result = runner.invoke(
+        cli.main, ["transliterate", "--from", "bundled", "Example", "a"]
+    )
+    assert bundled_result.exit_code == 0
+    assert bundled_result.output.strip() == "A"
+
+    # test multiple inputs with python output
+    bundled_multiple_result = runner.invoke(
+        cli.main, ["transliterate", "--from", "bundled", "Example", "a", "a"]
+    )
+    assert bundled_multiple_result.exit_code == 0
+    assert bundled_multiple_result.output.strip() == str(["A", "A"])
+
+    # test json
+    bundled_multiple_json_result = runner.invoke(
+        cli.main,
+        ["transliterate", "--from", "bundled", "Example", "--to", "json", "a", "a"],
+    )
+    assert bundled_multiple_json_result.exit_code == 0
+    assert bundled_multiple_json_result.output.strip() == json.dumps(["A", "A"])
+
+    # test transliterate from JSON
+    json_ = transliterator.dumps()
+    json_result = runner.invoke(
+        cli.main, ["transliterate", "--from", "json", json_, "a"]
+    )
+    assert json_result.exit_code == 0
+    assert json_result.output.strip() == "A"
+
+    # test transliterate from json file
+    json_file = tmpdir.mkdir("sub").join("test.json")
+    json_file.write(json_)
+    json_file_result = runner.invoke(
+        cli.main, ["transliterate", "--from", "json_file", json_file.strpath, "a"]
+    )
+    assert json_file_result.exit_code == 0
+    assert json_file_result.output.strip() == "A"
+
+    # test transliterate from yaml file
+    yaml_file = tmpdir.join("test.yaml")
+    yaml_file.write(test_yaml)
+    yaml_file_result = runner.invoke(
+        cli.main, ["transliterate", "--from", "yaml_file", yaml_file.strpath, "a"]
+    )
+    assert yaml_file_result.exit_code == 0
+    assert yaml_file_result.output.strip() == "A"
+
+
+def test_cli_generate_tests():
+    """Test tests command."""
+    runner = CliRunner()
+    gen_tests_result = runner.invoke(
+        cli.main, ["generate-tests", "--from", "bundled", "Example"]
+    )
+    assert gen_tests_result.exit_code == 0
+    yaml_ = yaml.safe_load(StringIO(gen_tests_result.output))
[…]
