### 2019-10-22 fd25ef0
update to paper
