### 2019-10-22 0dddce7
issues with bash kernel on RTD
