### 2019-10-22 92d6ccc
fixes to docs, etc.
