+++ docs/cli.rst
@@ -1,55 +1,98 @@
 
+.. -------------------------------------------------------------------------------------
+.. Note:
+..     This is a documentation source file for Graph Transliterator.
+..     Certain links and other features will not be accessible from here.
+.. Links:
+..     - Documentation: https://graphtransliterator.readthedocs.org
+..     - PyPI: https://pypi.org/project/graphtransliterator/
+..     - Repository: https://github.com/seanpue/graphtransliterator/
+.. -------------------------------------------------------------------------------------
+
+.. jupyter-execute::
+  :hide-code:
+
+  import graphtransliterator
+  import graphtransliterator.cli as cli
+  from click.testing import CliRunner
+  runner = CliRunner()
+  def run(func, parameters):
+    print(runner.invoke(func, parameters).output)
+
 ======================
 Command Line Interface
 ======================
-Graph Transliterator has a simple command line interface with four commands:
-``dump``, ``generate-tests``, ``test``, and ``transliterate``.
+Graph Transliterator has a simple command line interface with five commands:
+``dump``, ``generate-tests``, ``make-json``, ``test``, and ``transliterate``.
 
+
+.. code-block:: console
+
+  $ graphtransliterator --help
 
 .. jupyter-execute::
+  :hide-code:
 
-  %%bash
-  graphtransliterator --help
-
+  run(cli.main, ['--help'])
 
 Dump
 ----
 
 The ``dump`` command will output the specified transliterator as JSON:
 
-.. jupyter-execute::
+.. code-block:: console
 
-  %%bash
-  graphtransliterator dump --help
+  $ graphtransliterator dump --help
+
+.. jupyter-execute::
+  :hide-code:
+
+  run(cli.dump, ['--help'])
+
 
 It require a ``--from`` or ``-f`` option with two arguments. The first argument
 specifies the format of the transliterator (`bundled` or `yaml_file`) and the
 second a parameter for that format (the name of the bundled transliterator or the name
 of a YAML file).
 
-To load a bundled transliterator, give its name, which will be in camel case:
+To load a bundled transliterator, used `bundled` as the first parameter and give its
+(class) name, which will be in CamelCase, as the second:
+
+.. code-block:: console
+
+  $ graphtransliterator dump --from bundled Example
 
 .. jupyter-execute::
+  :hide-code:
 
-  %%bash
-  graphtransliterator dump --from bundled Example
+  run(cli.dump, ['--from', 'bundled', 'Example'])
 
-To load from a yaml file, give the name of the file:
+
+
+To load from a YAML file, give `yaml_file` as the first and the the name of the file as
+the second parameter:
+
+.. code-block:: console
+
+  $ graphtransliterator dump --from yaml_file ../graphtransliterator/transliterators/example/example.yaml
 
 
 .. jupyter-execute::
+  :hide-code:
 
-  %%bash
-  graphtransliterator dump --from yaml_file ../graphtransliterator/transliterators/example/example.yaml
+  run(cli.dump, ['--from', 'yaml_file', '../graphtransliterator/transliterators/example/example.yaml'])
 
 If you want to check for ambiguity in the transliterator before the dump, use the
 --`check-ambiguity` or `-ca` option:
 
+.. code-block:: console
+
+  $ graphtransliterator dump --from bundled Example --check-ambiguity
 
 .. jupyter-execute::
+  :hide-code:
 
-  %%bash
-  graphtransliterator dump --from bundled Example --check-ambiguity
+  run(cli.dump, ['--from', 'bundled', 'Example', '--check-ambiguity'])
 
 
 Generate Tests
[…]
