+++ docs/kudos.rst
@@ -1,0 +1,20 @@
+Kudos
+=====
+
+Graph Transliterator's developers acknowledge the following open-access projects, which
+have been particularly helpful in Graph Transliterator's development. These include:
+astropy_ (guide for documentation style expanding on numpy_), contributor_covenant_
+(basis for the code of conduct), `cookiecutter-pypackage` (initial Python module
+template), jupyter-sphinx_ (excellent renderer of live code results in the docs), and
+`marshmallow` (fantastic object serializer/deserializer and general model of good
+coding).
+
+Those from which code/text has been adopted are mentioned in `NOTICE`_.
+
+.. _NOTICE: https://github.com/seanpue/graphtransliterator/NOTICE
+.. _astropy: https://docs.astropy.org/en/stable/development/docrules.html
+.. _numpy: https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard
+.. _contributor_covenant: https://contributor-covenant.org
+.. _cookiecutter_pypackage: https://github.com/audreyr/cookiecutter-pypackage
+.. _jupyter-sphinx: https://github.com/jupyter/jupyter-sphinx
+.. _marshmallow: https://github.com/marshmallow-code/marshmallow
