+++ docs/bundled.rst
@@ -12,6 +12,9 @@
 Bundled Transliterators
 =======================
 
+.. note::
+<<Python code: :jupyter-download:script:`bundled`; Jupyter Notebook: :jupyter-download:notebook:`bundled`>>
+
 Graph Transliterator includes bundled transliterators in a :class:`Bundled` subclass of
 :class:`GraphTransliterator` that can be used as follows:
 
