+++ setup.py
@@ -50,6 +50,6 @@ setup(
     test_suite="tests",
     tests_require=test_requirements,
     url="https://github.com/seanpue/graphtransliterator",
-    version="0.3.8",
+    version="0.4.0",
     zip_safe=False,
 )
