+++ docs/conf.py
@@ -191,7 +191,7 @@ texinfo_documents = [
 
 # jupyter-sphinx SettingsSchema
 
-# jupyter_sphinx_linenos = True
+jupyter_sphinx_linenos = True
 jupyter_sphinx_continue_linenos = True
 ##def setup(app):
 ##   app.add_stylesheet('custom.css')
