+++ docs/usage.rst
@@ -12,7 +12,10 @@
 Usage
 =====
 
-<<Download code on this page here: :jupyter-download:script:`usage`>>
+
+.. note:
+
+  Python code on this page: :jupyter-download:script:`usage` Jupyter Notebook: :jupyter-download:notebook:`usage`
 
 To use Graph Transliterator in a project:
 
