+++ docs/tutorial.rst
@@ -11,8 +11,9 @@
 Tutorial: Using GraphTransliterator
 ===================================
 
-<<Download code on this page: :jupyter-download:script:`tutorial` or as a Jupyter
-Notebook: :jupyter-download:notebook:`tutorial` >>
+.. note::
+
+  Python code on this page: :jupyter-download:script:`tutorial` Jupyter Notebook: :jupyter-download:notebook:`tutorial`
 
 Graph Transliterator is designed to allow you to quickly develop rules for
 transliterating between languages and scripts. In this tutorial you will use a
