+++ README.rst
@@ -130,7 +130,7 @@ Sample Code and Graph
         version: "1.0.0"
   """).transliterate("hi")
 
-.. code-block:: none
+.. code-block:: python
 
     '¡ᴉɥ!'
 
