+++ requirements.txt
@@ -1,2 +1,3 @@
+Click
 marshmallow==3.2.0
 pyyaml==5.1.2
