### 2019-10-24 b9986ea
Pin click to latest version 7.0
