+++ requirements.txt
@@ -1,3 +1,3 @@
-Click
+Click==7.0
 marshmallow==3.2.0
 pyyaml==5.1.2
