### 2019-10-24 c2ea777
updated requirements_dev.txt
