+++ requirements.txt
@@ -1,3 +1,3 @@
-Click
-marshmallow==3.2.0
+Click==7.0
+marshmallow==3.2.1
 pyyaml==5.1.2
