### 2019-10-24 #1717 @seanpue
Hi @vc1492a @rlskoeser @gkthiruvathukal

Ok, I think it's ready to go. I have added the statement of need, updated the article, and also added bundled transliterators, a tutorial on adding bundled transliterators, and a command line interface. I have used jupyter-sphinx so that the docs automatically generate the sample code, and it can be cut and pasted or downloaded.
