### 2019-10-24 4fc1833
updates/fixes to docs
