+++ graphtransliterator/transliterators/itrans_devanagari_to_unicode/schemas.py
@@ -1,0 +1,236 @@
+# -*- coding: utf-8 -*-
+
+from marshmallow import (
+    fields,
+    post_load,
+    post_dump,
+    Schema,
+    validate,
+    ValidationError,
+    validates_schema,
+)
+
+from collections import defaultdict
+from .graphs import DirectedGraph
+from .initialize import _onmatch_rule_of, _transliteration_rule_of, _whitespace_rules_of
+from .process import RULE_RE, ONMATCH_RE
+from .rules import TransliterationRule
+import copy
+
+
+class WhitespaceDictSettingsSchema(Schema):
+    """Schema for Whitespace definition as a `dict`."""
+
+    default = fields.Str(required=True)
+    token_class = fields.Str(required=True)
+    consolidate = fields.Boolean(required=True)
+
+
+class WhitespaceSettingsSchema(WhitespaceDictSettingsSchema):
+    """Schema for Whitespace definition that loads as :class:`WhitespaceRules.`"""
+
+    @post_load
+    def make_whitespace_rules(self, data, **kwargs):
+        return _whitespace_rules_of(data)
+
+
+class EasyReadingSettingsSchema(Schema):
+    """Schema for easy reading settings.
+
+    Provides initial validation based on easy reading format.
+    """
+
+    tokens = fields.Dict(
+        keys=fields.Str(), values=fields.List(fields.Str()), required=True
+    )
+    rules = fields.Dict(
+        keys=fields.Str(validate=validate.Regexp(RULE_RE)),
+        values=fields.Str(),
+        required=True,
+    )
+    onmatch_rules = fields.List(
+        fields.Dict(
+            keys=fields.Str(validate=validate.Regexp(ONMATCH_RE)),
+            values=fields.Str(),
+            required=False,
+        )
+    )
+    metadata = fields.Dict(
+        keys=fields.Str(),
+        # no restriction on values
+        required=False,
+    )
+    whitespace = fields.Nested(WhitespaceDictSettingsSchema)
+
+
+class TransliterationRuleSchema(Schema):
+    """Schema for :class:`TransliterationRule`."""
+
+    production = fields.Str(required=True)
+    tokens = fields.List(fields.Str(), required=True)
+    prev_classes = fields.List(fields.Str(), allow_none=True)
+    next_classes = fields.List(fields.Str(), allow_none=True)
+    prev_tokens = fields.List(fields.Str(), allow_none=True)
+    next_tokens = fields.List(fields.Str(), allow_none=True)
+
+    class Meta:
+        fields = (
+            "production",
+            "prev_classes",
+            "prev_tokens",
+            "tokens",
+            "next_classes",
+            "next_tokens",
+            "cost",
+        )
+        ordered = True
+
+    @post_load
+    def make_transliteration_rule(self, data, **kwargs):
+        return _transliteration_rule_of(data)
+
+
+class OnMatchRuleSchema(Schema):
+    """Schema for :class:`OnMatchRule`."""
+
+    prev_classes = fields.List(fields.Str())
+    next_classes = fields.List(fields.Str())
+    production = fields.Str()
+
+    class Meta:
+        fields = ("prev_classes", "next_classes", "production")
+        ordered = True
+
+    @post_load
+    def make_onmatch_rule(self, data, **kwargs):
+        return _onmatch_rule_of(data)
+
+
+class SettingsSchema(Schema):
+    """Schema for settings in dictionary format.
+
+    Performs validation.
+    """
+
+    tokens = fields.Dict(
+        keys=fields.Str(), values=fields.List(fields.Str()), required=True
+    )
+    rules = fields.Nested(TransliterationRuleSchema, many=True, required=True)
[…]
