+++ docs/index.rst
@@ -10,10 +10,11 @@
 
 
 
+.. include:: ../README.rst
+
 .. toctree::
    :maxdepth: 2
 
-   about
    installation
    usage
    bundled
