+++ HISTORY.rst
@@ -265,4 +265,4 @@ History
 
 0.1.0 (2019-05-30)
 ------------------
-* First release on PyPI.
+* First release on PyPI. 
