### 2019-10-31 142fa07
updated test for make-json
