### 2019-10-31 2b65332
Bump version: 0.4.4 → 0.4.5
