+++ .travis.yml
@@ -11,7 +11,7 @@ deploy:
   distributions: sdist bdist_wheel
   user: seanpue
   password:
-    secure: tf8TO3ysOLNzuhXm4hcchgZYDs0RfbEhhDj0goOVNHOWjz7B/av3wgOKp5DG/xGy4tU7mqk74hrJia4SqSkqp+tuPP0yM2kYHcoCr4EhvdRxCM9yMmc7pj29sTwDBLXRlxEYpySMmBum5er1VZ/N4TRkgfCZoUdsogRI/ER2BQ3FJuiZkVEGECoZUnbdXiStRzh786qF3p9cSk7QY4g8WTYhxXLYQBFRSnioiRCHVcseDsXCoy2uXoyDffLzl5X5e6ad84hTJETKzFqcSZzLm0wkonkBswihC+sRcSRJKZVfTitCTSU6p7ns0AYB2wpRIIemsQOdjkreChaAITVQkx/f/Ml2wyp2egdXCxhVmBJmIqEw1nHn9IeTnwmMTuBuxuSuPEJr6IXaSitft0JzLMCgEiO3w0OhrlW0DcqL+s7SIiDnCK7xo5yezeYVb8M2pvP7tOoB52tBwkDBeRcASTxEltAQBWC/5WKiGpfsz4Au6vU9PwdfwZ/sL2oH0RVr1P+8ywv+jZmARmi8IPVcbbsOWg2QGSHYodbohFBCpajhAwI+LAGS+dberrhCM3TsFO8h5sBsrPNlo/JS/C0Wm8xaar6rdKXLN/oWgBM6HE3NI4E3/kH8LvZWqoRwRdJ0vvBRpvkOVXXj2GT/R6w4lEwgo4SRKWwkdLXXKbJU+tM=
+    secure: 1lQkfASXBF2WSN5EMnOzpAGlwaVDL9Rm+lxdFv3cKk07+wC/+PPfowAKRC0MrV3u9AU0w+IMXJpQc5DYS6/Etl4TTrbIkm68b5c7vHS2Gd8dpNjuD8tiGwOVX0g6vCLnWEHW0kLSxlxcVOLzzXU81rs8FH8YGr69+2GDkZye4dgviBHbSdD3VCysXhDSYzaOEuxfCVS/2tcsbKYgUc1v1u2acf0XYciZFFboJGH8EpCo1mliYqvTfb0rcx7TmGfyQq0lrvI6s3y0aaNIH92dtuprNpI6dtLfziYx2otbZXNkbiTLFg875TYiS5gadJu2XlRxvlqRYpwpTDDXHdmjIjw6/9WYHpdx0cRRcZlFx46peYOKPSgW72pAdy22c5UZ2tkEB/DmbQH7lyjX8XXfEEt8L5XU8n649laH1gCPhYI0R2dytAJpa2LrOyGCgc1LtR7Dryy0MgtOcRS6mKhqP2rfr1DX3++9vIu/T0vy3xYE4HsB8cGEyMI0VMyadWYm0EP+iCOp6dF1iDAUZiZ59tdlC3Q+jrJwZCXmxO4OVzYaZIsZ6Uu4BsEsCJWfNuOiKCysviZ4vu1M1TfWySv5PAWWNa8m2ZzeeUHCr9IfM2dNbRypZzFYZN/jv+qeU5g/68Dz7NYCgw/xGRjEjm0v6w/HEOYx/rgurHOYcs7t7Ls=
   on:
     repo: seanpue/graphtransliterator
     python: 3.7
