### 2019-10-31 fb417b6
updated user for travis.yml
