+++ .travis.yml
@@ -9,9 +9,9 @@ script: tox
 deploy:
   provider: pypi
   distributions: sdist bdist_wheel
-  user: "__token__"
+  user: __token__
   password:
-    secure: 1lQkfASXBF2WSN5EMnOzpAGlwaVDL9Rm+lxdFv3cKk07+wC/+PPfowAKRC0MrV3u9AU0w+IMXJpQc5DYS6/Etl4TTrbIkm68b5c7vHS2Gd8dpNjuD8tiGwOVX0g6vCLnWEHW0kLSxlxcVOLzzXU81rs8FH8YGr69+2GDkZye4dgviBHbSdD3VCysXhDSYzaOEuxfCVS/2tcsbKYgUc1v1u2acf0XYciZFFboJGH8EpCo1mliYqvTfb0rcx7TmGfyQq0lrvI6s3y0aaNIH92dtuprNpI6dtLfziYx2otbZXNkbiTLFg875TYiS5gadJu2XlRxvlqRYpwpTDDXHdmjIjw6/9WYHpdx0cRRcZlFx46peYOKPSgW72pAdy22c5UZ2tkEB/DmbQH7lyjX8XXfEEt8L5XU8n649laH1gCPhYI0R2dytAJpa2LrOyGCgc1LtR7Dryy0MgtOcRS6mKhqP2rfr1DX3++9vIu/T0vy3xYE4HsB8cGEyMI0VMyadWYm0EP+iCOp6dF1iDAUZiZ59tdlC3Q+jrJwZCXmxO4OVzYaZIsZ6Uu4BsEsCJWfNuOiKCysviZ4vu1M1TfWySv5PAWWNa8m2ZzeeUHCr9IfM2dNbRypZzFYZN/jv+qeU5g/68Dz7NYCgw/xGRjEjm0v6w/HEOYx/rgurHOYcs7t7Ls=
+    secure: 4oLUAb/fqlIAAxv+WnymJx3q7xwrsC8VTz/XiQEj0wWRkLt9q4tyC3zEUw3M+/FgYTuatwbOLX/eWkP7x/o3L19gm5c7PhvJ1lgFOj5bqT32u9Jw++POrC7ykKhmDkVElA4r5M5g5WgSfW2QZ9O0Ycj33UowJrMXi2UeBlCz72JRnA4o6zOTCbhdOqTSkGxrQhmIi5P3SFEgAhwT2mFeQjH2+XJOQxC4erFJQv299h81CyejooDKUGUDnUmFVlMmgGyyDXqYlo6bvzkmqwhffo3hLR6PfGs2q88kEryBozEZYG+7zrJ9AYiStL5n+VYeOxTkXKUiZyLDoRkl0j3Gr+KlFvyQU1mUT3dun143c161CkIpKNIL3f4DdXcjepB/um+tcsl04zlSiIgriMg7CH9fglXEOmL8VQxKvPebVRrLzuH88SS3IywUtZUR8jHnNrf+hDCSAGis80eQBBxlEkcZ0FdSQBmGZH2iEZHzTTLAXYbKdTGfiryZgF5Ym460KX4Xp9dqcgaGdI8yFHd8WsHKRNlx3g8XRe5w2upK4Rd63CtSjHSg6g0CfIcNSvHa1Q0I85oWs/Jvq0aM/5qqdHCpVnrHBMq6G79Zchzf9N1KZVooTbbQ42jRvhsmoWiH41SSEzX3O3rnPvRgWw/BGMlZ07A8uBs2f65sQMNRprs=
   on:
     repo: seanpue/graphtransliterator
     python: 3.7
