### 2019-11-04 d7d0cd2
updates to requirements and travis
