+++ requirements.txt
@@ -1,3 +1,3 @@
 Click==7.0
-marshmallow==3.2.1
+marshmallow==3.2.2
 pyyaml==5.1.2
