+++ .travis.yml
@@ -11,7 +11,7 @@ deploy:
   distributions: sdist bdist_wheel
   user: __token__
   password:
-    secure: 4oLUAb/fqlIAAxv+WnymJx3q7xwrsC8VTz/XiQEj0wWRkLt9q4tyC3zEUw3M+/FgYTuatwbOLX/eWkP7x/o3L19gm5c7PhvJ1lgFOj5bqT32u9Jw++POrC7ykKhmDkVElA4r5M5g5WgSfW2QZ9O0Ycj33UowJrMXi2UeBlCz72JRnA4o6zOTCbhdOqTSkGxrQhmIi5P3SFEgAhwT2mFeQjH2+XJOQxC4erFJQv299h81CyejooDKUGUDnUmFVlMmgGyyDXqYlo6bvzkmqwhffo3hLR6PfGs2q88kEryBozEZYG+7zrJ9AYiStL5n+VYeOxTkXKUiZyLDoRkl0j3Gr+KlFvyQU1mUT3dun143c161CkIpKNIL3f4DdXcjepB/um+tcsl04zlSiIgriMg7CH9fglXEOmL8VQxKvPebVRrLzuH88SS3IywUtZUR8jHnNrf+hDCSAGis80eQBBxlEkcZ0FdSQBmGZH2iEZHzTTLAXYbKdTGfiryZgF5Ym460KX4Xp9dqcgaGdI8yFHd8WsHKRNlx3g8XRe5w2upK4Rd63CtSjHSg6g0CfIcNSvHa1Q0I85oWs/Jvq0aM/5qqdHCpVnrHBMq6G79Zchzf9N1KZVooTbbQ42jRvhsmoWiH41SSEzX3O3rnPvRgWw/BGMlZ07A8uBs2f65sQMNRprs=
+    secure: atulVDMRLEUHKGrD3Mtr8j+o3wr3SUed1OwA17ubAIWiAPB7JnRO/klDxvH/Ss1x4tW0jxrSt4YWQ3MuMhd8qWif3Szdi1Ta9oW969Jy91NX5xLIy8NOR6joPVWSI+V9sGulmrTPdGlcty7dbe05MYNxjhJCsD/8kulGZ0spmxAfPRiAoNu9YlB2mE+ySCAZwf3oNJGgAEXedWD3T3eO/1WSd7JSxLHamt4xKgAR2OiNVVhHGT+OcXWzab+VRsp95MamgBYAgzNYehJNvN0jeGFnPXV3ZFYe01eCyAUREDfaQeM4UScf/4tnY3AAmpVhoF/4FioU5bBGHSE0JPyXT9SkpzRLsT6XyZv9+pu/7bxkFkssaR6PbpVUQVlOW+AtcMCBBaNOQWKaFzu3AIOTTER19AgT014CtCR90JwYh8HcRvVJwDWli5UwRsrBwU48+lYwhUbjJcxSth+//0Sc7FXfeZ4ZJoU+wcKUDlgHGXklIM6bo/U8Sgc2z7uAZaAZzOTRNwhwcmmptowbSKim7/NhL9AN0uyNyDr+UzAItIStMLKsh5/TIRsQi02FgAcEGkPN7ZmGW2KPGL7rn8CwQDd5ZnGT3wjsbJkls3wfoByb3GrEqmLpnwnZkMOEB9TIjzhz6WHYGidVa/Wowy2tul1JPKeADY0x4OIsu7jkR/g=
   on:
     repo: seanpue/graphtransliterator
     python: 3.7
