+++ setup.py
@@ -22,8 +22,6 @@ setup(
         "Intended Audience :: Developers",
         "License :: OSI Approved :: MIT License",
         "Natural Language :: English",
-        "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.5",
         "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
     ],
@@ -45,7 +43,7 @@ setup(
             'graphtransliterator=graphtransliterator.cli:main',
         ],
     },
-    python_requires=">=3.5",
+    python_requires=">=3.6",
     setup_requires=setup_requirements,
     test_suite="tests",
     tests_require=test_requirements,
