+++ HISTORY.rst
@@ -13,6 +13,11 @@ History
 * add static typing with mypy
 
 
+0.4.7 (2019-11-04)
+------------------
+* temp switch back to dev version of jupyter-sphinx for overflow error
+* updated twine to 2.0 so dropped Python 3.5 support
+
 0.4.6 (2019-11-04)
 ------------------
 * switched to latest jupyter-sphinx
