+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py35, py36, py37, flake8
+envlist = py36, py37, flake8
 
 [flake8]
 max-line-length = 88
@@ -8,7 +8,6 @@ max-line-length = 88
 python =
     3.7: py37
     3.6: py36
-    3.5: py35
 
 [testenv:flake8]
 basepython = python
