### 2019-11-08 #1717 @seanpue
Hi @vc1492a @rlskoeser @gkthiruvathukal Wanted to check to see if you've had a chance to review the updates. Thanks.
