### 2019-11-08 #1717 @gkthiruvathukal
@seanpue It looks like we are nearing a conclusion here! It would be great if  @vc1492a and @rlskoeser can chime in and take a final look. If I don't hear any new issues, say, by Monday, I think we are  ready to move toward acceptance.
