### 2019-11-08 #1717 @vc1492a
@seanpue must have missed the notification previously, apologies! I took a look at the repository and the updates look good to me. I feel that this work is ready for publication in JOSS.
