### 2019-11-21 #1717 @seanpue
Thanks again @vc1492a  @rlskoeser Do you think you'll get a chance to review the changes?  I believe I have addressed your concerns.
