### 2019-11-21 #1717 @rlskoeser
I apologize for the delay. Changes look good to me.
