+++ setup.py
@@ -18,7 +18,7 @@ setup(
     author="A. Sean Pue",
     author_email="pue@msu.edu",
     classifiers=[
-        "Development Status :: 4 - Beta",
+        "Development Status :: 5 - Production/Stable",
         "Intended Audience :: Developers",
         "License :: OSI Approved :: MIT License",
         "Natural Language :: English",
