+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.4.10
+current_version = 1.0.0
 commit = True
 tag = True
 
