### 2019-11-26 #1717 @gkthiruvathukal
@seanpue Thanks so much to you and our wonderful reviewers @rlskoeser and @vc1492a.

@openjournals/joss-eics I am ready to move on accepting this contribution!
