### 2019-11-29 #1717 @Kevin-Mattheus-Moerman
@seanpue this work is about to be processed for acceptance. Can you give the paper a final check yourself as well (in particular the author names and affiliations)?
