### 2019-11-29 #1717 @Kevin-Mattheus-Moerman
@seanpue once you fixed/checked the above can you post an archived version of the software on Zenodo and report back the DOI of the archived version?
