### 2019-11-30 eaf7d90
Revert "Bump version: 1.0.1 → 1.0.2"
This reverts commit e889b7e38224cc33fd776e19b5c2eba168f9e3b4.
