+++ README.rst
@@ -22,6 +22,9 @@ Graph Transliterator
 .. image:: https://img.shields.io/pypi/pyversions/graphtransliterator
        :alt: PyPI - Python Version
 
+ .. image:: https://zenodo.org/badge/189464608.svg
+    :target: https://zenodo.org/badge/latestdoi/189464608
+
 A graph-based transliteration tool that lets you convert the symbols of one
 language or script to those of another using rules that you define.
 
