+++ paper/paper.bib
@@ -23,7 +23,8 @@
   number={6},
   pages={93},
   year={2018},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s12046-018-0828-8}
 }
 
 @inproceedings{durrani2010hindi,
@@ -63,7 +64,8 @@
   author={Chen, Nancy and Banchs, Rafael E and Zhang, Min and Duan, Xiangyu and Li, Haizhou},
   booktitle={Proceedings of the Seventh Named Entities Workshop},
   pages={55--73},
-  year={2018}
+  year={2018},
+  doi={10.18653/v1/w18-2409}
 }
 @Manual{YAMLAintMarkup,
   title = {{{YAML~Ain}}'t~{{Markup~Language}}~({{YAML}}\texttrademark{}) {{Version}}~1.2},
