### 2019-11-30 #1717 @Kevin-Mattheus-Moerman
@seanpue can you please change the Zenodo archive metadata such that the title and author list matches your paper?
Can you confirm the version is now v1.0.1?
