### 2019-11-30 79dca5f
minor paper upddate
