+++ README.rst
@@ -22,8 +22,8 @@ Graph Transliterator
 .. image:: https://img.shields.io/pypi/pyversions/graphtransliterator
      :alt: PyPI - Python Version
 
- .. image:: https://zenodo.org/badge/189464608.svg
-      :target: https://zenodo.org/badge/latestdoi/189464608
+.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.3558365.svg
+     :target: https://doi.org/10.5281/zenodo.3558365
 
 A graph-based transliteration tool that lets you convert the symbols of one
 language or script to those of another using rules that you define.
