### 2019-11-30 0bf9004
updates to docs
