### 2019-11-30 #1717 @seanpue
@Kevin-Mattheus-Moerman I have fixed the Zenodo metadata. Please update the version to 1.0.4 Thanks!
