### 2019-12-01 #1717 @Kevin-Mattheus-Moerman
> @Kevin-Mattheus-Moerman The PDF is not showing my initials in the footer, ...

This is normal
