### 2019-12-01 #1717 @Kevin-Mattheus-Moerman
@openjournals/dev  Can you check why this DOI is still not resolving?
