### 2019-12-02 #1717 @danielskatz
This DOI resolves, so I'm going to close this issue
Thanks to @gkthiruvathukal for editing and @rlskoeser & @vc1492a for reviewing, and congratulations to @seanpue
